### 2023-08-14 #129 @bluegenes
`sourmash` submission
**Submitting Author:** Tessa Pierce-Ward (@bluegenes)
**All current maintainers:** @ctb, @luizirber, @bluegenes
**Package Name:** sourmash
**One-Line Description of Package:** `sourmash` is a command line tool and Python library for sketching collections of DNA, RNA, and amino acid k-mers for biological sequence search, comparison, and analysis.
**Repository Link:** https://github.com/sourmash-bio/sourmash
**Version submitted:** 4.8
**Editor:** @snacktavish Emily Jane McTavish
**EiC**: @NickleDave
**Reviewer 1:** @LilyAnderssonLee Lili Andersson-Li
**Reviewer 2:** @elais Elais Player
**Archive:** [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10951577.svg)](https://doi.org/10.5281/zenodo.10951577)
**Version accepted:** [v4.8.8](https://github.com/sourmash-bio/sourmash/releases/tag/v4.8.8)
**Date accepted (month/day/year):** 4/4/2024

---

> Large collections of genomes, transcriptomes, and raw sequencing data sets are readily available in biology. With the scale of data now available, the field needs lightweight computational methods for searching and summarizing the content of both public and private collections. `sourmash` implements FracMinHash sketching, a lightweight technique that supports accurate estimation of overlap and containment between two sequencing data sets. `sourmash` provides a flexible set of programmatic functionality for sequence search and comparison, together with a robust and well-tested command-line interface.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

> The target audience for sourmash is biologists looking to compare biological sequencing datasets of any kind. `sourmash`'s FracMinHash sketching produces a small, irreversible representation of each dataset, allowing users to search and compare without compromising private data. We provide a user-friendly CLI with mature functionality as well as a python API and experimental Rust API for computational biologists with advanced use cases.

>There are a number of sketching tools for genomic data (e.g. [Mash](https://github.com/marbl/Mash), [kmindex](https://github.com/tlemane/kmindex), [Dashing](https://github.com/dnbaker/dashing)), each of which implements a slightly different sketching technique. `sourmash`'s FracMinHash enables accurate comparisons of sets of very different sizes (unlike standard MinHash implemented in, e.g. Mash), enabling analysis of genomes contained within metagenomes ([Irber et al. 2022](https://www.biorxiv.org/content/10.1101/2022.01.11.475838v2)), which supports taxonomic profiling ([Portik et al., 2022](https://doi.org/10.1186/s12859-022-05103-0)) and [content-based search of publicly-available metagenomes](https://dib-lab.github.io/2022-paper-branchwater-software/). `sourmash` now additionally supports estimation of Average Nucleotide Identity from FracMinHash ([Rahman Hera et al., 2023](https://doi.org/10.1101/gr.277651.123)). The sourmash team has focused on maintaining a robust user interface and continually improving functionality and user experience, with ~ monthly software releases.

 > N/A

> in 'developer notes', not main README.

- [x] The package is deposited in a long-term repository with the DOI: 10.21105/joss.00027
