### 2023-08-15 c35cc19
MRG: change `matching_hashes.sig` to `matching-hashes.sig` in documentation (#2713)
Fixes https://github.com/sourmash-bio/sourmash/issues/2709
Fixes https://github.com/sourmash-bio/sourmash/issues/2708

For review -
- [ ] updated [prefetch
section](https://sourmash--2713.org.readthedocs.build/en/2713/command-line.html#caveats-and-comments)
- [ ] updated [sig intersect
docs](https://sourmash--2713.org.readthedocs.build/en/2713/command-line.html#sourmash-signature-intersect-intersect-two-or-more-signatures)
- [ ] updated [sig inflate
docs](https://sourmash--2713.org.readthedocs.build/en/2713/com
[…]
