### 2023-08-16 6c28366
[MRG] update JOSS paper for v4 (#1361)
Per https://github.com/openjournals/joss/issues/423, we are planning to
make a new submission to JOSS for sourmash v4.x. The first paper from
2016 is [here](https://joss.theoj.org/papers/10.21105/joss.00027).

Fixes #1321.

Ref https://github.com/dib-lab/sourmash/issues/622,
https://github.com/dib-lab/sourmash/issues/444.

- [x] update authors per https://github.com/dib-lab/sourmash/issues/1367
- [x] fix `@@@` in paper.md YAML header
- [x] review [JOSS
guidelines](https://joss.readthedocs.io/en/latest/submitting.html) and
think about what to add, if anyth
[…]
