+++ pyproject.toml
@@ -13,6 +13,7 @@ version = "4.8.3-dev"
 
 authors = [
   { name="Luiz Irber", orcid="0000-0003-4371-9659" },
+  { name="N. Tessa Pierce-Ward", orcid="0000-0002-2942-5331" },
   { name="Mohamed Abuelanin", orcid="0000-0002-3419-4785" },
   { name="Harriet Alexander", orcid="0000-0003-1308-8008" },
   { name="Abhishek Anant", orcid="0000-0002-5751-2010" },
@@ -33,7 +34,6 @@ authors = [
   { name="Marisa Lim", orcid="0000-0003-2097-8818" },
   { name="Ricky Lim", orcid="0000-0003-1313-7076" },
   { name="Ivan Ogasawara", orcid="0000-0001-5049-4289" },
-  { name="N. Tessa Pierce", orcid="0000-0002-2942-5331" },
   { name="Taylor Reiter", orcid="0000-0002-7388-421X" },
   { name="Camille Scott", orcid="0000-0001-8822-8779" },
   { name="Andreas Sjödin", orcid="0000-0001-5350-4219" },
