### 2023-08-19 8a54cc5
add tests for detecting overwrite
