### 2023-08-19 5fcd12a
fix filename == '-' issue
