+++ tests/test_sourmash.py
@@ -3874,7 +3874,10 @@ def test_multigather_metagenome_output(runtmp):
     cmd = cmd.split(' ')
     c.run_sourmash(*cmd)
 
-    output_csv = runtmp.output('-.csv') # @CTB
+    print(runtmp.last_result.out)
+    print(runtmp.last_result.err)
+
+    output_csv = runtmp.output('b92dbf45dd57867cbec2321ccfa55af8.csv')
     assert os.path.exists(output_csv)
     with open(output_csv, newline='') as fp:
         x = fp.readlines()
@@ -3904,7 +3907,7 @@ def test_multigather_metagenome_output_outdir(runtmp):
     cmd = cmd.split(' ')
     c.run_sourmash(*cmd)
 
-    output_csv = runtmp.output('savehere/-.csv') # @CTB
+    output_csv = runtmp.output('savehere/b92dbf45dd57867cbec2321ccfa55af8.csv')
     assert os.path.exists(output_csv)
     with open(output_csv, newline='') as fp:
         x = fp.readlines()
@@ -5106,7 +5109,7 @@ def test_multigather_output_unassigned_with_abundance(runtmp):
     assert "the recovered matches hit 91.0% of the abundance-weighted query." in out
     assert "the recovered matches hit 57.2% of the query k-mers (unweighted)." in out
 
-    assert os.path.exists(c.output('r3.fa.unassigned.sig')) # @CTB
+    assert os.path.exists(c.output('r3.fa.unassigned.sig'))
 
     nomatch = sourmash.load_one_signature(c.output('r3.fa.unassigned.sig'))
     assert nomatch.minhash.track_abundance
