+++ .github/workflows/rust.yml
@@ -63,17 +63,6 @@ jobs:
           toolchain: ${{ matrix.rust }}
           override: true
 
-      - name: Set up Python 3.8
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.8"
-
-      - name: Install dependencies
-        continue-on-error: ${{ matrix.continue }}
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e .
-
       - name: Run tests
         uses: actions-rs/cargo@v1
         with:
@@ -89,16 +78,6 @@ jobs:
         with:
           toolchain: stable
           override: true
-
-      - name: Set up Python 3.8
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.8"
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e .
 
       - uses: actions-rs/install@v0.1
         with:
@@ -254,6 +233,12 @@ jobs:
           profile: minimal
           toolchain: stable
           override: true
+
+      - name: Check semver
+        uses: obi1kenobi/cargo-semver-checks-action@v2
+        with:
+          crate-name: sourmash
+          version-tag-prefix: r
 
       - name: Make sure we can publish the sourmash crate
         uses: actions-rs/cargo@v1
