+++ .github/workflows/dev_envs.yml
@@ -57,7 +57,7 @@ jobs:
 
     - name: install dependencies
       shell: bash -l {0}
-      run: mamba install tox-conda rust git compilers pandoc
+      run: mamba install tox>=4.2 tox-conda rust git compilers pandoc
 
     - name: run tests for 3.9
       shell: bash -l {0}
