+++ tox.ini
@@ -1,5 +1,7 @@
 [tox]
-envlist =
+requires =
+  tox>=4.2
+env_list =
   py310,
   py311,
   py39,
@@ -11,20 +13,19 @@ envlist =
   hypothesis,
   khmer,
   khmer_master
-minversion = 3.12
-isolated_build = true
+minversion = 4.2
 skip_missing_interpreters = true
 
 [testenv]
+description = run the tests with pytest under {basepython}
 package = wheel
 wheel_build_env = .pkg
-description = run the tests with pytest under {basepython}
-setenv =
+set_env =
     PIP_DISABLE_VERSION_CHECK = 1
     COVERAGE_FILE = {env:COVERAGE_FILE:{toxworkdir}/.coverage.{envname}}
     VIRTUALENV_NO_DOWNLOAD = 1
     PIP_EXTRA_INDEX_URL = https://antocuni.github.io/pypy-wheels/manylinux2010
-passenv =
+pass_env =
     TOXENV
     CURL_CA_BUNDLE
     http_proxy
@@ -36,6 +37,8 @@ passenv =
     PIP_CACHE_DIR
     CI
     PYTHONDEVMODE
+    LIBCLANG_PATH
+    BINDGEN_EXTRA_CLANG_ARGS
 deps =
     pip >= 19.3.1
 extras =
@@ -103,8 +106,8 @@ description = invoke sphinx-build to build the HTML docs
 basepython = python3.10
 extras = doc
 whitelist_externals = pandoc
-passenv = HOME
-changedir = {toxinidir}
+pass_env = HOME
+change_dir = {toxinidir}
 #commands = sphinx-build -d "{toxworkdir}/docs_doctree" doc "{toxworkdir}/docs_out" --color -W -bhtml {posargs}
 commands = sphinx-build -d "{toxworkdir}/docs_doctree" doc "{toxworkdir}/docs_out" --color -bhtml {posargs}
            python -c 'import pathlib; print("documentation available under file://\{0\}".format(pathlib.Path(r"{toxworkdir}") / "docs_out" / "index.html"))'
@@ -117,7 +120,7 @@ deps = twine >= 1.12.1
        readme-renderer[md] >= 24.0
        pip >= 19.1
 skip_install = true
-changedir = {toxinidir}
+change_dir = {toxinidir}
 extras =
 commands = pip wheel -w {envtmpdir}/build --no-deps .
            twine check {envtmpdir}/build/*
@@ -125,7 +128,7 @@ commands = pip wheel -w {envtmpdir}/build --no-deps .
 [testenv:mypy]
 description = run mypy checker
 basepython = python3.8
-passenv = {[testenv]passenv}
+pass_env = {[testenv]pass_env}
           # without PROGRAMDATA cloning using git for Windows will fail with an `error setting certificate verify locations` error
           PROGRAMDATA
 deps  = mypy
@@ -134,7 +137,7 @@ commands = mypy src/sourmash
 [testenv:fix_lint]
 description = format the code base to adhere to our styles, and complain about what we cannot do automatically
 basepython = python3.8
-passenv = {[testenv]passenv}
+pass_env = {[testenv]pass_env}
           # without PROGRAMDATA cloning using git for Windows will fail with an `error setting certificate verify locations` error
           PROGRAMDATA
           PRE_COMMIT_HOME
@@ -151,9 +154,9 @@ deps = {[testenv]deps}
        coverage >= 5.0.1
        diff_cover
 skip_install = True
-passenv = {[testenv]passenv}
+pass_env = {[testenv]pass_env}
           DIFF_AGAINST
-setenv = COVERAGE_FILE={toxworkdir}/.coverage
[…]
