### 2023-08-19 c1a7825
stick with tox<4 for now, tox-conda is not compatible
