+++ tox.ini
@@ -1,6 +1,4 @@
 [tox]
-requires =
-  tox>=4.2
 env_list =
   py310,
   py311,
@@ -13,7 +11,8 @@ env_list =
   hypothesis,
   khmer,
   khmer_master
-minversion = 4.2
+min_version = 3.27
+isolated_build = true
 skip_missing_interpreters = true
 
 [testenv]
