### 2023-08-19 0dd0e3b
build: updates spun off mastiff PR (#2723)
(There are too many things happening in #2230, so I'll open a couple of
PRs to simplify it)

- Bring tox config as close as possible to `tox 4`, but keep using `tox
3` until [`tox-conda` is compatible with `tox
4`](https://github.com/tox-dev/tox-conda/issues/156). Ended up adding an
explicit `'tox>=3.27,<4'` to the conda dev env instructions to get the
right version, and remove/replace deprecated options from `tox.ini`
- Bump `MACOSX_DEPLOYMENT_TARGET` to `11.0` consistently for maturin and
CI.
- Remove Python installation for Rust tests in CI (not us
[…]
