### 2023-08-20 d35a61d
build: remove specific tooling that breaks s390x
