+++ pyproject.toml
@@ -178,17 +178,6 @@ build-verbosity = 3
 CARGO_REGISTRIES_CRATES_IO_PROTOCOL="sparse"
 PATH="$HOME/.cargo/bin:$PATH"
 
-[tool.cibuildwheel.linux]
-before-all = [
-  "curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain=stable",
-  "cargo update --dry-run",
-  "if [ -f /etc/system-release ]; then yum -y install centos-release-scl; fi",
-  "if [ -f /etc/system-release ]; then yum -y install llvm-toolset-7.0; fi",
-]
-before-build = [
-  "if [ -f /etc/system-release ]; then source scl_source enable llvm-toolset-7.0; fi",
-  "if [ -f /etc/system-release ]; then source scl_source enable devtoolset-10; fi",
-]
 [tool.cibuildwheel.linux.environment]
 CARGO_REGISTRIES_CRATES_IO_PROTOCOL="sparse"
 PATH="$HOME/.cargo/bin:$PATH"
