### 2023-08-20 710b48d
[skip ci] Start implementing FFI for manifest, picklist and selection
