+++ src/sourmash/picklist.py
@@ -298,6 +298,24 @@ class SignaturePicklist:
             if self.__contains__(ss):
                 yield ss
 
+    def _as_rust(self):
+        from ._lowlevel import ffi, lib
+        from .utils import rustcall, decode_str
+
+        ptr = lib.picklist_new()
+
+        rustcall(lib.picklist_set_coltype, ptr, self.coltype.encode('utf-8'), len(self.coltype))
+        rustcall(lib.picklist_set_pickfile, ptr, self.pickfile.encode('utf-8'), len(self.pickfile))
+        rustcall(lib.picklist_set_column_name, ptr, self.column_name.encode('utf-8'), len(self.column_name))
+        rustcall(lib.picklist_set_pickstyle, ptr, self.pickstyle.value)
+
+        #self.preprocess_fn = preprocess[coltype]
+        #self.pickset = None
+        #self.found = set()
+        #self.n_queries = 0
+
+        return ptr
+
 
 def passes_all_picklists(ss, picklists):
     "does the signature 'ss' pass all of the picklists?"
