+++ src/sourmash/index/__init__.py
@@ -34,10 +34,14 @@ signatures on demand. No signatures are kept in memory.
 CounterGather - an ancillary class returned by the 'counter_gather()' method.
 """
 
+from __future__ import annotations
+
 import os
 import sourmash
 from abc import abstractmethod, ABC
-from collections import namedtuple, Counter
+from collections import defaultdict, Counter
+from typing import NamedTuple, Optional, TypedDict, TYPE_CHECKING
+import weakref
 
 from sourmash.search import (make_jaccard_search_query,
                              make_containment_query,
@@ -45,12 +49,78 @@ from sourmash.search import (make_jaccard_search_query,
 from sourmash.manifest import CollectionManifest
 from sourmash.logging import debug_literal
 from sourmash.signature import load_signatures, save_signatures
+from sourmash._lowlevel import ffi, lib
+from sourmash.utils import RustObject, rustcall, decode_str, encode_str
+from sourmash import SourmashSignature
+from sourmash.picklist import SignaturePicklist
 from sourmash.minhash import (flatten_and_downsample_scaled,
                               flatten_and_downsample_num,
                               flatten_and_intersect_scaled)
 
-# generic return tuple for Index.search and Index.gather
-IndexSearchResult = namedtuple('Result', 'score, signature, location')
+if TYPE_CHECKING:
+    from typing_extensions import Unpack
+
+
+class IndexSearchResult(NamedTuple):
+    """generic return tuple for Index.search and Index.gather"""
+    score: float
+    signature: SourmashSignature
+    location: str
+
+
+class Selection(TypedDict):
+    ksize: Optional[int]
+    moltype: Optional[str]
+    num: Optional[int]
+    scaled: Optional[int]
+    containment: Optional[bool]
+    abund: Optional[bool]
+    picklist: Optional[SignaturePicklist]
+
+
+# TypedDict can't have methods (it is a dict in runtime)
+def _selection_as_rust(selection: Selection):
+    ptr = lib.selection_new()
+
+    for key, v in selection.items():
+        if v is not None:
+            if key == "ksize":
+                rustcall(lib.selection_set_ksize, ptr, v)
+
+            elif key == "moltype":
+                hash_function = None
+                if v.lower() == "dna":
+                    hash_function = lib.HASH_FUNCTIONS_MURMUR64_DNA
+                elif v.lower() == "protein":
+                    hash_function = lib.HASH_FUNCTIONS_MURMUR64_PROTEIN
+                elif v.lower() == "dayhoff":
+                    hash_function = lib.HASH_FUNCTIONS_MURMUR64_DAYHOFF
+                elif v.lower() == "hp":
+                    hash_function = lib.HASH_FUNCTIONS_MURMUR64_HP
+
+                rustcall(lib.selection_set_moltype, ptr, hash_function)
+
+            elif key == "num":
+                rustcall(lib.selection_set_num, ptr, v)
+
+            elif key == "scaled":
+                rustcall(lib.selection_set_scaled, ptr, v)
+
+            elif key ==  "containment":
+                rustcall(lib.selection_set_containment, ptr, v)
+
+            elif key == "abund":
+                rustcall(lib.selection_set_abund, ptr, bool(v))
+
+            elif key == "picklist":
+                picklist_ptr = v._as_rust()
+                rustcall(lib.selection_set_picklist, ptr, picklist_ptr)
+
+            else:
+                raise KeyError(f"Unsupported key {key} for Selection in rust")
+
+    return ptr
+
 
 class Index(ABC):
     # this will be removed soon; see sourmash#1894.
@@ -307,8 +377,7 @@ class Index(ABC):
         return counter
 
     @abstractmethod
[…]

[…]
