### 2023-08-20 c09d2ec
[skip ci] FFI for Rust LinearIndex
