+++ src/core/src/ffi/minhash.rs
@@ -6,12 +6,33 @@ use crate::encodings::{aa_to_dayhoff, aa_to_hp, translate_codon, HashFunctions};
 use crate::ffi::utils::{ForeignObject, SourmashStr};
 use crate::signature::SeqToHashes;
 use crate::signature::SigsTrait;
-use crate::sketch::minhash::KmerMinHash;
+use crate::sketch::hyperloglog::HyperLogLog;
+use crate::sketch::minhash::{
+    AbundMinHashOps, FracMinHashOps, KmerMinHash, KmerMinHashBTree, MinHashOps,
+};
+use crate::sketch::Sketch;
+use crate::Error;
+use crate::HashIntoType;
+
+#[derive(Clone)]
+pub enum MinHash {
+    Mutable(KmerMinHashBTree),
+    Frozen(KmerMinHash),
+}
 
 pub struct SourmashKmerMinHash;
 
+#[no_mangle]
+pub unsafe extern "C" fn kmerminhash_to_mutable(
+    ptr: *const SourmashKmerMinHash,
+) -> *mut SourmashKmerMinHash {
+    let mh = SourmashKmerMinHash::as_rust(ptr);
+
+    SourmashKmerMinHash::from_rust(mh.clone().to_mutable())
+}
+
 impl ForeignObject for SourmashKmerMinHash {
-    type RustObject = KmerMinHash;
+    type RustObject = MinHash;
 }
 
 #[no_mangle]
@@ -23,14 +44,23 @@ pub unsafe extern "C" fn kmerminhash_new(
     track_abundance: bool,
     n: u32,
 ) -> *mut SourmashKmerMinHash {
-    let mh = KmerMinHash::new(scaled, k, hash_function, seed, track_abundance, n);
+    let mh = KmerMinHashBTree::new(scaled, k, hash_function, seed, track_abundance, n);
 
-    SourmashKmerMinHash::from_rust(mh)
+    SourmashKmerMinHash::from_rust(MinHash::Mutable(mh))
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn kmerminhash_free(ptr: *mut SourmashKmerMinHash) {
     SourmashKmerMinHash::drop(ptr);
+}
+
+#[no_mangle]
+pub unsafe extern "C" fn kmerminhash_to_frozen(
+    ptr: *const SourmashKmerMinHash,
+) -> *mut SourmashKmerMinHash {
+    let mh = SourmashKmerMinHash::as_rust(ptr);
+
+    SourmashKmerMinHash::from_rust(mh.clone().to_frozen())
 }
 
 #[no_mangle]
@@ -471,6 +501,7 @@ unsafe fn kmerminhash_similarity(ptr: *const SourmashKmerMinHash, other: *const 
     mh.similarity(other_mh, ignore_abundance, downsample)
 }
 }
+
 ffi_fn! {
 unsafe fn kmerminhash_angular_similarity(ptr: *const SourmashKmerMinHash, other: *const SourmashKmerMinHash)
                                          -> Result<f64> {
@@ -478,4 +509,373 @@ unsafe fn kmerminhash_angular_similarity(ptr: *const SourmashKmerMinHash, other:
     let other_mh = SourmashKmerMinHash::as_rust(other);
     mh.angular_similarity(other_mh)
 }
+}
+
+impl MinHash {
+    pub fn to_mutable(self) -> MinHash {
+        match self {
+            MinHash::Mutable(mh) => MinHash::Mutable(mh),
+            MinHash::Frozen(mh) => MinHash::Mutable(mh.into()),
+        }
+    }
+
+    pub fn to_frozen(self) -> MinHash {
+        match self {
[…]
