+++ src/core/src/signature.rs
@@ -28,6 +28,16 @@ pub trait SigsTrait {
     fn seed(&self) -> u64;
 
     fn hash_function(&self) -> HashFunctions;
+    fn set_hash_function(&mut self, h: HashFunctions) -> Result<(), Error>;
+    fn is_protein(&self) -> bool {
+        self.hash_function() == HashFunctions::murmur64_protein
+    }
+    fn dayhoff(&self) -> bool {
+        self.hash_function() == HashFunctions::murmur64_dayhoff
+    }
+    fn hp(&self) -> bool {
+        self.hash_function() == HashFunctions::murmur64_hp
+    }
 
     fn add_hash(&mut self, hash: HashIntoType);
 
@@ -114,6 +124,14 @@ impl SigsTrait for Sketch {
             Sketch::MinHash(ref mh) => mh.hash_function(),
             Sketch::LargeMinHash(ref mh) => mh.hash_function(),
             Sketch::HyperLogLog(ref hll) => hll.hash_function(),
+        }
+    }
+
+    fn set_hash_function(&mut self, h: HashFunctions) -> Result<(), Error> {
+        match *self {
+            Sketch::MinHash(ref mut mh) => mh.set_hash_function(h),
+            Sketch::LargeMinHash(ref mut mh) => mh.set_hash_function(h),
+            Sketch::HyperLogLog(ref mut hll) => hll.set_hash_function(h),
         }
     }
 
@@ -683,6 +701,8 @@ impl Default for Signature {
 
 impl PartialEq for Signature {
     fn eq(&self, other: &Signature) -> bool {
+        use crate::sketch::minhash::{KmerMinHash, KmerMinHashBTree};
+
         let metadata = self.class == other.class
             && self.email == other.email
             && self.hash_function == other.hash_function
@@ -691,14 +711,25 @@ impl PartialEq for Signature {
 
         // TODO: find the right signature
         // as long as we have a matching
-        if let Sketch::MinHash(mh) = &self.signatures[0] {
-            if let Sketch::MinHash(other_mh) = &other.signatures[0] {
-                return metadata && (mh == other_mh);
-            }
-        } else {
-            unimplemented!()
+        match &self.signatures[0] {
+            Sketch::MinHash(mh) => match &other.signatures[0] {
+                Sketch::MinHash(other_mh) => metadata && (mh == other_mh),
+                Sketch::LargeMinHash(other_mh) => {
+                    // TODO: avoid clone
+                    metadata && (mh == &Into::<KmerMinHash>::into(other_mh.clone()))
+                }
+                Sketch::HyperLogLog(_) => todo!(),
+            },
+            Sketch::LargeMinHash(mh) => match &other.signatures[0] {
+                Sketch::LargeMinHash(other_mh) => metadata && (mh == other_mh),
+                Sketch::MinHash(other_mh) => {
+                    // TODO: avoid clone
+                    metadata && (mh == &Into::<KmerMinHashBTree>::into(other_mh.clone()))
+                }
+                Sketch::HyperLogLog(_) => todo!(),
+            },
+            Sketch::HyperLogLog(_) => todo!(),
         }
-        metadata
     }
 }
 
