### 2023-08-21 323d2c3
MRG: update #2065 (uniquify CSV output from multigather) (#2721)
Note: PR into #2065.

This PR updates #2065 with all the changes from `latest`. This includes
the fix & update to multigather in
https://github.com/sourmash-bio/sourmash/pull/2322, which:
- fixes the output of multigather to include more than one line 😆 
- prints out the output filename
- supports `--output-dir`

No functional changes are made beyond the merge, so some tests are still
failing; will discuss fixes in yet a new PR :).

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Keya Barve <5332849
[…]
