+++ paper.bib
@@ -1,4 +1,4 @@
-@article{ondov2015fast,
+@article{Ondov:2015,
   title={Fast genome and metagenome distance estimation using MinHash},
   author={Ondov, Brian D and Treangen, Todd J and Mallonee, Adam B and Bergman, Nicholas H and Koren, Sergey and Phillippy, Adam M},
   journal={bioRxiv},
@@ -7,4 +7,63 @@
   publisher={Cold Spring Harbor Labs Journals},
   doi={10.1101/029827},
   url={https://doi.org/10.1101/029827}
+}
+
+@article{Brown:2016,
+  doi = {10.21105/joss.00027},
+  url = {https://doi.org/10.21105/joss.00027},
+  year = {2016},
+  publisher = {The Open Journal},
+  volume = {1},
+  number = {5},
+  pages = {27},
+  author = {C. Titus Brown and Luiz Irber},
+  title = {sourmash: a library for MinHash sketching of DNA},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Pierce:2019,
+  doi = {10.12688/f1000research.19675.1},
+  url = {https://doi.org/10.12688/f1000research.19675.1},
+  year = {2019},
+  month = jul,
+  publisher = {F1000 Research Ltd},
+  volume = {8},
+  pages = {1006},
+  author = {N. Tessa Pierce and Luiz Irber and Taylor Reiter and Phillip Brooks and C. Titus Brown},
+  title = {Large-scale sequence comparisons with sourmash},
+  journal = {F1000Research}
+}
+@article{gather,
+  title={Lightweight compositional analysis of metagenomes with FracMinHash and minimum metagenome covers},
+  author={Irber, Luiz Carlos and Brooks, Phillip T and Reiter, Taylor E and Pierce-Ward, N Tessa and Hera, Mahmudur Rahman and Koslicki, David and Brown, C Titus},
+  journal={bioRxiv},
+  year={2022},
+  publisher={Cold Spring Harbor Laboratory}
+}
+
+@article{branchwater,
+  title={Sourmash Branchwater Enables Lightweight Petabyte-Scale Sequence Search},
+  author={Irber, Luiz Carlos and Pierce-Ward, N Tessa and Brown, C Titus},
+  journal={bioRxiv},
+  year={2022},
+  publisher={Cold Spring Harbor Laboratory}
+}
+
+@article{koslicki2019improving,
+  title={Improving minhash via the containment index with applications to metagenomic analysis},
+  author={Koslicki, David and Zabeti, Hooman},
+  journal={Applied Mathematics and Computation},
+  volume={354},
+  pages={206--215},
+  year={2019},
+  publisher={Elsevier}
+}
+
+@article{hera2022debiasing,
+  title={Debiasing FracMinHash and deriving confidence intervals for mutation rates across a wide range of evolutionary distances},
+  author={Hera, Mahmudur Rahman and Pierce-Ward, N Tessa and Koslicki, David},
+  journal={bioRxiv},
+  year={2022},
+  publisher={Cold Spring Harbor Laboratory}
 }
