+++ src/sourmash/tax/__main__.py
@@ -4,15 +4,16 @@ Command-line entry point for 'python -m sourmash.tax'
 import sys
 import csv
 import os
-from collections import defaultdict
+from collections import defaultdict, Counter
+from dataclasses import asdict, fields
+import re
 
 import sourmash
-from ..sourmash_args import FileOutputCSV
-from sourmash.logging import set_quiet, error, notify
-from sourmash.lca.lca_utils import display_lineage
+from ..sourmash_args import FileOutputCSV, FileInputCSV, FileOutput
+from sourmash.logging import set_quiet, error, notify, print_results
 
 from . import tax_utils
-from .tax_utils import ClassificationResult, MultiLineageDB
+from .tax_utils import MultiLineageDB, RankLineageInfo, LINLineageInfo, AnnotateTaxResult
 
 usage='''
 sourmash taxonomy <command> [<args>] - manipulate/work with taxonomy information.
@@ -31,27 +32,32 @@ metagenome -g <gather_csv> [<gather_csv> ... ] -t [<taxonomy_csv> ...]    - summ
 sourmash taxonomy metagenome -h
 '''
 
-# some utils
+# outfile utils
+_output_type_to_ext = {
+    'csv_summary': '.summarized.csv',
+    'classification': '.classifications.csv',
+    'krona': '.krona.tsv',
+    'lineage_summary': '.lineage_summary.tsv',
+    'annotate': '.with-lineages.csv',
+    'human': '.human.txt',
+    'lineage_csv': '.lineage.csv',
+    'kreport': ".kreport.txt",
+    'lingroup': ".lingroup.tsv",
+    'bioboxes': '.bioboxes.profile'
+    }
+
 def make_outfile(base, output_type, *, output_dir = ""):
     limit_float_decimals=False
     if base == "-":
         limit_float_decimals=True
         return base, limit_float_decimals
-    ext=""
-    if output_type == 'csv_summary':
-        ext = '.summarized.csv'
-    elif output_type == 'classification':
-        ext = '.classifications.csv'
-    elif output_type == 'krona':
-        ext = '.krona.tsv'
-    elif output_type == 'lineage_summary':
-        ext = '.lineage_summary.tsv'
-    elif output_type == 'annotate':
-        ext = '.with-lineages.csv'
+
+    ext = _output_type_to_ext[output_type]
+
     fname = base+ext
     if output_dir:
         fname = os.path.join(output_dir, fname)
-    notify(f"saving `{output_type}` output to {fname}.")
+    notify(f"saving '{output_type}' output to '{fname}'.")
     return fname, limit_float_decimals
 
 
@@ -67,7 +73,7 @@ def metagenome(args):
         tax_assign = MultiLineageDB.load(args.taxonomy_csv,
                        keep_full_identifiers=args.keep_full_identifiers,
                        keep_identifier_versions=args.keep_identifier_versions,
-                       force=args.force)
+                       force=args.force, lins=args.lins)
         available_ranks = tax_assign.available_ranks
     except ValueError as exc:
         error(f"ERROR: {str(exc)}")
@@ -84,53 +90,105 @@ def metagenome(args):
     # next, collect and load gather results
     gather_csvs = tax_utils.collect_gather_csvs(args.gather_csv, from_file= args.from_file)
     try:
-        gather_results, idents_missed, total_missed, _ = tax_utils.check_and_load_gather_csvs(gather_csvs, tax_assign, force=args.force,
-                                                                                       fail_on_missing_taxonomy=args.fail_on_missing_taxonomy)
+         query_gather_results = tax_utils.check_and_load_gather_csvs(gather_csvs, tax_assign, force=args.force,
+                                                                     fail_on_missing_taxonomy=args.fail_on_missing_taxonomy,
+                                                                     keep_full_identifiers=args.keep_full_identifiers,
[…]

[…]
