### 2023-08-21 3d8da8e
Bump typed-builder from 0.14.0 to 0.15.2
Bumps [typed-builder](https://github.com/idanarye/rust-typed-builder) from 0.14.0 to 0.15.2.
- [Changelog](https://github.com/idanarye/rust-typed-builder/blob/master/CHANGELOG.md)
- [Commits](https://github.com/idanarye/rust-typed-builder/commits)

---
updated-dependencies:
- dependency-name: typed-builder
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
