### 2023-08-22 d6dd054
add new option + tests + docs
