### 2023-08-23 53e7241
more cleanup of tests
