### 2023-08-23 3d0467e
test -E/--extension
