+++ src/core/Cargo.toml
@@ -8,10 +8,10 @@ keywords = ["minhash", "bioinformatics"]
 categories = ["science", "algorithms", "data-structures"]
 license = "BSD-3-Clause"
 edition = "2021"
-rust-version = "1.56.1"
 readme = "README.md"
 autoexamples = false
 autobins = false
+rust-version = "1.64.0"
 
 [lib]
 name = "sourmash"
@@ -29,45 +29,45 @@ az = "1.0.0"
 bytecount = "0.6.0"
 byteorder = "1.4.3"
 cfg-if = "1.0"
-counter = "0.5.4"
-finch = { version = "0.4.1", optional = true }
+counter = "0.5.7"
+finch = { version = "0.6.0", optional = true }
 fixedbitset = "0.4.0"
+getrandom = { version = "0.2", features = ["js"] }
 getset = "0.1.1"
-log = "0.4.8"
+log = "0.4.20"
 md5 = "0.7.0"
 murmurhash3 = "0.0.5"
 niffler = { version = "2.3.1", default-features = false, features = [ "gz" ] }
 nohash-hasher = "0.2.0"
 num-iter = "0.1.43"
-once_cell = "1.3.1"
-rayon = { version = "1.5.2", optional = true }
-serde = { version = "1.0.137", features = ["derive"] }
-serde_json = "1.0.80"
+once_cell = "1.18.0"
+rayon = { version = "1.7.0", optional = true }
+serde = { version = "1.0.168", features = ["derive"] }
+serde_json = "1.0.105"
 primal-check = "0.3.1"
 thiserror = "1.0"
-typed-builder = "0.10.0"
+typed-builder = "0.14.0"
 twox-hash = "1.6.0"
-vec-collections = "0.3.4"
-piz = "0.4.0"
-memmap2 = "0.5.0"
-ouroboros = "0.15.0"
+vec-collections = "0.4.3"
+piz = "0.5.0"
+memmap2 = "0.7.1"
+ouroboros = "0.17.2"
 
 rkyv = { version = "0.7.39", optional = true }
 rocksdb = { version = "0.21.0", optional = true }
 flume = "0.10.13"
-roaring = "0.9.0"
+roaring = "0.10.0"
 histogram = "0.6.9"
 numsep = "0.1.12"
 size = "0.4.0"
 
 [dev-dependencies]
 assert_matches = "1.3.0"
-criterion = "0.3.2"
-needletail = { version = "0.4.1", default-features = false }
-proptest = { version = "1.0.0", default-features = false, features = ["std"]}
+criterion = "0.5.1"
+needletail = { version = "0.5.1", default-features = false }
+proptest = { version = "1.2.0", default-features = false, features = ["std"]}
 rand = "0.8.2"
-getrandom = { version = "0.2", features = ["js"] }
-tempfile = "3.1.0"
+tempfile = "3.7.0"
 
 [[bench]]
 name = "index"
