+++ src/core/Cargo.toml
@@ -83,7 +83,7 @@ version = "0.3.64"
 features = ["console", "File"]
 
 [target.'cfg(all(target_arch = "wasm32"))'.dependencies.chrono]
-version = "0.4.26"
+version = "0.4.28"
 features = ["wasmbind"]
 
 [target.'cfg(all(target_arch = "wasm32", target_os="unknown"))'.dev-dependencies]
