### 2023-09-05 038a37b
Build: Remove deprecated system_packages from readthedocs (#2740)
`python.system_packages` is deprecated and now breaks readthedocs build.
Turns out we have all the dependencies properly specified, so no other
changes needed.

More info: https://blog.readthedocs.com/drop-support-system-packages/
