### 2023-09-06 eadc56b
simplify picklist API
