### 2023-09-07 d900c1d
apply fix to 'manifest.to_picklist()' fn
