### 2023-09-10 f6bd61f
fix SqliteCollectionManifest.to_picklist
