### 2023-09-10 ffb42c8
add grep test from #2602
