### 2023-09-11 bb24f32
Bump chrono from 0.4.26 to 0.4.30 (#2755)
Bumps [chrono](https://github.com/chronotope/chrono) from 0.4.26 to
0.4.30.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/chronotope/chrono/releases">chrono's
releases</a>.</em></p>
<blockquote>
<h2>0.4.30</h2>
<p>In this release, we have decided to swap out the
<code>chrono::Duration</code> type (which has been a re-export of time
0.1 <code>Duration</code> type) with our own definition, which exposes a
strict superset of the <code>time::Duration</code> API. This helps avoid
warnings about the <a
href="https
[…]
