### 2023-09-11 1e2853f
Bump ouroboros from 0.17.2 to 0.18.0 (#2743)
Bumps [ouroboros](https://github.com/joshua-maros/ouroboros) from 0.17.2
to 0.18.0.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/joshua-maros/ouroboros/commits">compare
view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibility
score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=ouroboros&package-manager=cargo&previous-version=0.17.2&new-version=0.18.0)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-updates#abo
[…]
