+++ .github/workflows/dev_envs.yml
@@ -11,7 +11,7 @@ jobs:
       with:
         fetch-depth: 0
 
-    - uses: cachix/install-nix-action@v22
+    - uses: cachix/install-nix-action@v23
       with:
         extra_nix_config: |
           access-tokens = github.com=${{ secrets.GITHUB_TOKEN }}
