### 2023-09-11 35e1171
Bump supercharge/redis-github-action from 1.6.0 to 1.7.0
Bumps [supercharge/redis-github-action](https://github.com/supercharge/redis-github-action) from 1.6.0 to 1.7.0.
- [Release notes](https://github.com/supercharge/redis-github-action/releases)
- [Changelog](https://github.com/supercharge/redis-github-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/supercharge/redis-github-action/compare/1.6.0...1.7.0)

---
updated-dependencies:
- dependency-name: supercharge/redis-github-action
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by
[…]
