+++ .github/workflows/build_wheel_all_archs.yml
@@ -35,7 +35,7 @@ jobs:
       fail-fast: false
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
         with:
           fetch-depth: 0
 
