### 2023-09-13 6df0360
MRG: fix duplicate md5 in picklist problem (#2747)
This PR changes picklists using `manifest`, `prefetch`, `search`, and
`gather` coltypes to use a composite tuple of `(ident, md5short)` for
selection. This should have essentially zero false positives, unlike the
current behavior (which relies solely on md5).

The basic problem solved: when two signatures with different signature
metadata (name, etc) but the same md5sum were in a single index,
grepping/extracting via metadata (e.g. with a name) would return all
signatures with matching md5sum.

This PR fixes this behavior, and also cleans up `
[…]
