### 2023-09-15 3632dd7
fix exponential behavior
