+++ src/sourmash/manifest.py
@@ -232,11 +232,11 @@ class CollectionManifest(BaseCollectionManifest):
         self._add_rows([row])
 
     def _add_rows(self, rows):
-        self.rows.extend(rows)
-
-        # maintain a fast check for md5sums for __contains__ check.
         md5set = self._md5_set
-        for row in self.rows:
+
+        # only iterate once, in case it's a generator
+        for row in rows:
+            self.rows.append(row)
             md5set.add(row['md5'])
 
     def __iadd__(self, other):
