+++ tests/test_manifest.py
@@ -1,6 +1,7 @@
 """
 Tests for manifest code in databases, etc.
 """
+import pytest
 from io import StringIO
 
 import sourmash
@@ -49,7 +50,8 @@ def test_manifest_operations():
         siglist.append(sig)
 
     manifest = index.CollectionManifest(rows)
-    manifest += manifest
+    manifest2 = index.CollectionManifest(rows)
+    manifest += manifest2
 
     assert len(manifest) == 2*len(rows)
     assert len(manifest) == 4
@@ -57,6 +59,24 @@ def test_manifest_operations():
     md5_list = [ row['md5'] for row in manifest.rows ]
     assert '16869d2c8a1d29d1c8e56f5c561e585e' in md5_list
     assert '120d311cc785cc9d0df9dc0646b2b857' in md5_list
+
+
+def test_manifest_operations_fail():
+    # test basic manifest operations - +=
+    protzip = utils.get_test_data('prot/protein.zip')
+
+    loader = sourmash.load_file_as_index(protzip)
+
+    rows = []
+    siglist = []
+    for (sig, loc) in loader._signatures_with_internal():
+        row = index.CollectionManifest.make_manifest_row(sig, loc)
+        rows.append(row)
+        siglist.append(sig)
+
+    manifest = index.CollectionManifest(rows)
+    with pytest.raises(Exception):
+        manifest += manifest
 
 
 def test_manifest_to_picklist():
@@ -167,6 +187,7 @@ def test_save_load_manifest():
 def test_manifest_to_picklist_bug(runtmp):
     # this tests a fun combination of things that led to a bug.
     # tl;dr we only want to iterate once across a generator...
+    # ref #2762
     c = runtmp
     all_zip = utils.get_test_data('prot/all.zip')
 
