### 2023-09-15 d13dd24
ensure iterate only once across rows being added
