### 2023-09-15 94d82b1
MRG: fix exponential time explosion in `sig check` (#2762)
Fixes #2646.

Embarrassing "let's make things run really slow" typo of the month...
fixed!

In brief, the following code in `manifest.py` was causing an exponential
explosion in time cost:
```python
    def _add_rows(self, rows):
        self.rows.extend(rows)

        # maintain a fast check for md5sums for __contains__ check.
        md5set = self._md5_set
        for row in self.rows:
            md5set.add(row['md5'])
```
when called many times on a single row, because each addition of a row
triggered an iteration over ALL rows.

T
[…]
