+++ include/sourmash.h
@@ -43,6 +43,7 @@ enum SourmashErrorCode {
   SOURMASH_ERROR_CODE_SERDE_ERROR = 100004,
   SOURMASH_ERROR_CODE_NIFFLER_ERROR = 100005,
   SOURMASH_ERROR_CODE_CSV_ERROR = 100006,
+  SOURMASH_ERROR_CODE_ROCKS_DB_ERROR = 100007,
 };
 typedef uint32_t SourmashErrorCode;
 
