### 2023-09-17 b77a452
make storage and manifest private in collection
