### 2023-09-17 6e69b7f
cleanup template, replace with selection
