+++ src/core/src/index/revindex/mem_revindex.rs
@@ -84,7 +84,7 @@ impl LinearIndex {
 impl RevIndex {
     pub fn new(
         search_sigs: &[PathBuf],
-        template: &Sketch,
+        selection: &Selection,
         threshold: usize,
         queries: Option<&[KmerMinHash]>,
         _keep_sigs: bool,
@@ -92,8 +92,7 @@ impl RevIndex {
         // If threshold is zero, let's merge all queries and save time later
         let merged_query = queries.and_then(|qs| Self::merge_queries(qs, threshold));
 
-        let collection =
-            Collection::from_paths(search_sigs)?.select(&Selection::from_template(template))?;
+        let collection = Collection::from_paths(search_sigs)?.select(&selection)?;
         let linear = LinearIndex::from_collection(collection.try_into()?);
 
         Ok(linear.index(threshold, merged_query, queries))
@@ -101,7 +100,7 @@ impl RevIndex {
 
     pub fn from_zipfile<P: AsRef<Path>>(
         zipfile: P,
-        template: &Sketch,
+        selection: &Selection,
         threshold: usize,
         queries: Option<&[KmerMinHash]>,
         _keep_sigs: bool,
@@ -109,8 +108,7 @@ impl RevIndex {
         // If threshold is zero, let's merge all queries and save time later
         let merged_query = queries.and_then(|qs| Self::merge_queries(qs, threshold));
 
-        let collection =
-            Collection::from_zipfile(zipfile)?.select(&Selection::from_template(template))?;
+        let collection = Collection::from_zipfile(zipfile)?.select(&selection)?;
         let linear = LinearIndex::from_collection(collection.try_into()?);
 
         Ok(linear.index(threshold, merged_query, queries))
@@ -130,15 +128,14 @@ impl RevIndex {
 
     pub fn new_with_sigs(
         search_sigs: Vec<Signature>,
-        template: &Sketch,
+        selection: &Selection,
         threshold: usize,
         queries: Option<&[KmerMinHash]>,
     ) -> Result<RevIndex> {
         // If threshold is zero, let's merge all queries and save time later
         let merged_query = queries.and_then(|qs| Self::merge_queries(qs, threshold));
 
-        let collection =
-            Collection::from_sigs(search_sigs)?.select(&Selection::from_template(template))?;
+        let collection = Collection::from_sigs(search_sigs)?.select(selection)?;
         let linear = LinearIndex::from_collection(collection.try_into()?);
 
         let idx = linear.index(threshold, merged_query, queries);
@@ -338,24 +335,18 @@ impl<'a> Index<'a> for RevIndex {
 mod test {
     use super::*;
 
+    use crate::index::revindex::prepare_query;
     use crate::sketch::minhash::max_hash_for_scaled;
     use crate::Result;
 
     #[test]
     fn revindex_new() -> Result<()> {
-        let max_hash = max_hash_for_scaled(10000);
-        let template = Sketch::MinHash(
-            KmerMinHash::builder()
-                .num(0u32)
-                .ksize(31)
-                .max_hash(max_hash)
-                .build(),
-        );
+        let selection = Selection::builder().ksize(31).scaled(10000).build();
         let search_sigs = [
             "../../tests/test-data/gather/GCF_000006945.2_ASM694v2_genomic.fna.gz.sig".into(),
             "../../tests/test-data/gather/GCF_000007545.1_ASM754v1_genomic.fna.gz.sig".into(),
         ];
-        let index = RevIndex::new(&search_sigs, &template, 0, None, false)?;
+        let index = RevIndex::new(&search_sigs, &selection, 0, None, false)?;
         assert_eq!(index.colors.len(), 3);
 
         Ok(())
@@ -363,21 +354,14 @@ mod test {
 
[…]
