### 2023-09-17 95bcd2e
Placate linters on HashFunctions
