### 2023-09-17 9bb513c
Initial support for custom hash function
