### 2023-09-18 7afebda
collection: build from_sigs and from_paths in parallel if possible
