+++ .github/workflows/rust.yml
@@ -109,7 +109,7 @@ jobs:
         run: cargo binstall --no-confirm cargo-tarpaulin
 
       - name: Coverage with tarpaulin
-        run: cargo tarpaulin --all --all-features --timeout 600 --out Xml -- --test-threads 1
+        run: cargo tarpaulin --all --all-features --timeout 600 --out xml -- --test-threads 1
 
       - name: Upload Rust coverage to codecov
         uses: codecov/codecov-action@v3
