+++ src/core/Cargo.toml
@@ -45,7 +45,7 @@ serde = { version = "1.0.168", features = ["derive"] }
 serde_json = "1.0.107"
 primal-check = "0.3.1"
 thiserror = "1.0"
-typed-builder = "0.14.0"
+typed-builder = "0.16.1"
 twox-hash = "1.6.0"
 vec-collections = "0.3.4"
 piz = "0.5.0"
