### 2023-09-19 019a944
Bump pypa/cibuildwheel from 2.15.0 to 2.16.0 (#2768)
Bumps [pypa/cibuildwheel](https://github.com/pypa/cibuildwheel) from
2.15.0 to 2.16.0.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/pypa/cibuildwheel/releases">pypa/cibuildwheel's
releases</a>.</em></p>
<blockquote>
<h2>v2.16.0</h2>
<ul>
<li>✨ Add the ability to pass additional flags to a build frontend
through the <a
href="https://cibuildwheel.readthedocs.io/en/stable/options/#build-frontend">CIBW_BUILD_FRONTEND</a>
option (<a
href="https://redirect.github.com/pypa/cibuildwheel/issues/1588">#158
[…]
