+++ doc/command-line.md
@@ -493,7 +493,7 @@ signatures, rather than all the signatures in the database.
 ## `sourmash tax` subcommands for integrating taxonomic information into gather results
 
 The `sourmash tax` subcommands support taxonomic analysis of genomes
-and taxonomic profiling of metagenomes.  
+and taxonomic profiling of metagenomes.
 See
 [taxonomic profiling with sourmash](classifying-signatures.md#taxonomic-profiling-with-sourmash)
 for more information.
@@ -1091,18 +1091,18 @@ and 120,757 within the `p__Proteobacteria`.
 These commands use LCA databases (created with `lca index`, below, or
 prepared databases such as [genbank-k31.lca.json.gz](databases.md)).
 
+### `sourmash lca classify` - classify a genome using an LCA database
+
+`sourmash lca classify` classifies one or more signatures using the given
+list of LCA DBs. It is meant for classifying metagenome-assembled genome
+bins (MAGs) and single-cell genomes (SAGs).
+
 ```{attention}
 We no longer recommend using `sourmash lca` for taxonomic analysis;
 please use `sourmash tax` instead.  See
 [taxonomic profiling with sourmash](classifying-signatures.md#taxonomic-profiling-with-sourmash)
 for more information.
 ```
-
-### `sourmash lca classify` - classify a genome using an LCA database
-
-`sourmash lca classify` classifies one or more signatures using the given
-list of LCA DBs. It is meant for classifying metagenome-assembled genome
-bins (MAGs) and single-cell genomes (SAGs).
 
 Usage:
 
@@ -1157,7 +1157,7 @@ a taxonomic disagreement, and the taxid & lineage refer to the name at
 that rank (the least-common-ancestor at which an assignment can be
 made).
 
-For example, if you saw this line in the CSV file: 
+For another example, if you saw this line in the CSV file: 
 
 ```
 TARA_ASW_MAG_00029,1224,disagree,phylum,Bacteria;Proteobacteria
@@ -1185,6 +1185,13 @@ exploring metagenomes and metagenome-assembled genome bins.
 `sourmash lca summarize` also weights output with hash abundances, so
 that output percentages are weighted by the number of times a k-mer is
 seen; this can be turned off with `--ignore-abundance`.
+
+```{attention}
+We no longer recommend using `sourmash lca` for taxonomic analysis;
+please use `sourmash tax` instead.  See
+[taxonomic profiling with sourmash](classifying-signatures.md#taxonomic-profiling-with-sourmash)
+for more information.
+```
 
 Usage:
 
