### 2023-09-25 256a705
MRG: add `__all__` to `sig/__main__.py` (#2778)
This change limits what is imported with `from sourmash.sig.__main__
import *`.

Fixes https://github.com/sourmash-bio/sourmash/issues/2714
