### 2023-09-25 a103dd3
add text about LCA remaining useful
