### 2023-09-25 66bc910
MRG: adjust documentation to recommend `tax` over `lca` for taxonomic analysis (#2777)
This PR adds cautionary notes to the command line docs, and updates the
information on classifying signatures to suggest using tax instead of
LCA, and even explains why :).

There is more work to be done - we need to add more tutorials, and
adjust the language in classifying-signatures around gather and LCA -
but this is a nice standalone PR!

Fixes https://github.com/sourmash-bio/sourmash/issues/2562
Fixes https://github.com/sourmash-bio/sourmash/issues/2772
Fixes https://github.com/sourmash-bio/sourmash/is
[…]
