+++ src/core/Cargo.toml
@@ -40,7 +40,7 @@ niffler = { version = "2.3.1", default-features = false, features = [ "gz" ] }
 nohash-hasher = "0.2.0"
 num-iter = "0.1.43"
 once_cell = "1.18.0"
-rayon = { version = "1.7.0", optional = true }
+rayon = { version = "1.8.0", optional = true }
 serde = { version = "1.0.168", features = ["derive"] }
 serde_json = "1.0.107"
 primal-check = "0.3.1"
