### 2023-09-25 10013af
minor updates to release procedure
