### 2023-09-26 24121d0
add usage str for 'sig ingest'
