+++ doc/faq.md
@@ -1,16 +1,22 @@
 # Frequently Asked Questions
 
+```{contents} Contents
+:depth: 3
+```
+
 ## How is sourmash different from mash?
 
-mash is an awesome piece of software that inspired sourmash - hence the
-name we chose for sourmash! But are they the same? No, they are not!
+[mash](https://mash.readthedocs.org/) is an awesome piece of software
+that inspired sourmash - hence the name we chose for sourmash! But are
+they the same? No, they are not!
 
-mash is based on MinHash sketching, a technique for randomly selecting
-a very small set of k-mers to represent a potentially very large
-sequence. MinHash sketches can be used to estimate Jaccard similarity
-(a distance metric that lets you find closely related sequences)
-with very high accuracy, under one condition - that the two sequences
-be of similar size.
+mash is based on
+[MinHash sketching](https://en.wikipedia.org/wiki/MinHash), a
+technique for randomly selecting a very small set of k-mers to
+represent a potentially very large sequence. MinHash sketches can be
+used to estimate Jaccard similarity (a distance metric that lets you
+find closely related sequences) with very high accuracy, under one
+condition - that the two sequences be of similar size.
 
 So, MinHash sketching is great when comparing bacterial genomes, which are
 all around 2-10 MB in size.
@@ -20,10 +26,10 @@ genomes, because metagenomes are usually _much_ larger than genomes.
 
 So we built sourmash around a different kind of sketch - FracMinHash -
 which can estimate Jaccard similarity between sets of very different
-sizes. FracMinHash sketches also support overlap and containment analysis,
-and are convenient in a variety of other ways - see our paper on
-FracMinHash, [Lightweight compositional analysis of metagenomes with
-FracMinHash and minimum metagenome covers](@@).
+sizes. FracMinHash sketches also support overlap and containment
+analysis, and are convenient in a variety of other ways - see our
+paper on FracMinHash,
+[Lightweight compositional analysis of metagenomes with FracMinHash and minimum metagenome covers](https://www.biorxiv.org/content/10.1101/2022.01.11.475838v2).
 
 There are some drawbacks to FracMinHash sketches - read on!
 
@@ -60,13 +66,87 @@ that match above the genus level - family or above.
 
 If you go higher (k=51), a higher percentage of k-mers are genome-specific.
 
-For the core sourmash operations - search, gather, and compare -
-we believe (with evidence!) that (a) the differences between k=21,
-k=31, and k=51 are negligible; and that (b) k=31 works fine for
-most day-to-day use of sourmash.
+For the core sourmash operations - search, gather, and compare - we
+believe (with evidence!) that (a) the differences between k=21, k=31,
+and k=51 are negligible; and that (b) k=31 works fine for most
+day-to-day use of sourmash.
 
-We also provide [Genbank and GTDB databases](@@) for k=21, k=31, and
-k=51, too.
+We also provide [Genbank and GTDB databases](databases.md) for k=21,
+k=31, and k=51.
+
+For some background on k-mer specificity, we recommend this paper:
+[MetaPalette: a k-mer Painting Approach for Metagenomic Taxonomic Profiling and Quantification of Novel Strain Variation](https://journals.asm.org/doi/10.1128/msystems.00020-16),
+Koslicki & Falush, 2016.
 
 ## How do k-mer-based analyses compare with read mapping?
 
+tl;dr very well! But it's a bit one sided: if k-mers match, reads will
+map, but not necessarily vice versa. So read mapping rates are almost always
+higher than k-mer matching rates.
+
+### Mapping reads to reference vs k-mer "detection" or containment
+
+Let's start by looking at a simple example: suppose you have Illumina
+shotgun sequencing of a new isolate, and you want to compare it to a
+reference genome for a member of the same species.  You calculate
[…]
