### 2023-09-26 1d5b754
update / add verify notes
