+++ src/core/Cargo.toml
@@ -46,7 +46,6 @@ murmurhash3 = "0.0.5"
 niffler = { version = "2.3.1", default-features = false, features = [ "gz" ] }
 nohash-hasher = "0.2.0"
 num-iter = "0.1.43"
-numsep = "0.1.12"
 once_cell = "1.18.0"
 ouroboros = "0.18.0"
 piz = "0.5.0"
@@ -56,14 +55,12 @@ rkyv = { version = "0.7.39", optional = true }
 roaring = "0.10.0"
 serde = { version = "1.0.168", features = ["derive"] }
 serde_json = "1.0.107"
-size = "0.4.0"
 thiserror = "1.0"
 twox-hash = "1.6.0"
 typed-builder = "0.14.0"
 vec-collections = "0.4.3"
 
 [dev-dependencies]
-assert_matches = "1.3.0"
 criterion = "0.5.1"
 needletail = { version = "0.5.1", default-features = false }
 proptest = { version = "1.2.0", default-features = false, features = ["std"]}
