### 2023-09-27 966c045
MRG: run rust checks always (#2787)
Fixes https://github.com/sourmash-bio/sourmash/issues/2550

There were some concerns about how long things take to run, so we'll
evaluate that and maybe revert this change if it slows down PRs.

Alternatively, we could implement #2342 and make better use of draft
PRs.
