+++ .github/workflows/rust.yml
@@ -4,11 +4,6 @@ on:
   push:
     branches: [latest]
   pull_request:
-    paths:
-      - 'Cargo.lock'
-      - 'src/core/**'
-      - 'tests/test-data/**'
-      - '.github/workflows/rust.yml'
   schedule:
     - cron: "0 0 * * *" # daily
 
