### 2023-09-29 2a3a284
cleanup and comment plot tests
