+++ tests/test_sourmash.py
@@ -838,8 +838,10 @@ def test_do_plot_comparison(runtmp):
     assert os.path.exists(c.output("cmp.matrix.png"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_2(c):
+def test_do_plot_comparison_2_pdf(runtmp):
+    # test plot --pdf
+    c = runtmp
+
     testdata1 = utils.get_test_data('short.fa')
     testdata2 = utils.get_test_data('short2.fa')
     c.run_sourmash('sketch', 'translate', '-p', 'k=31,num=500', testdata1, testdata2)
@@ -851,8 +853,10 @@ def test_do_plot_comparison_2(c):
     assert os.path.exists(c.output("cmp.matrix.pdf"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_3(c):
+def test_do_plot_comparison_3(runtmp):
+    # test plot --labels
+    c = runtmp
+
     testdata1 = utils.get_test_data('short.fa')
     testdata2 = utils.get_test_data('short2.fa')
     c.run_sourmash('sketch', 'translate', '-p', 'k=31,num=500', testdata1, testdata2)
@@ -865,8 +869,10 @@ def test_do_plot_comparison_3(c):
     assert os.path.exists(c.output("cmp.matrix.png"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_4_output_dir(c):
+def test_do_plot_comparison_4_output_dir(runtmp):
+    # test plot --output-dir
+    c = runtmp
+
     output_dir = c.output('xyz_test')
 
     testdata1 = utils.get_test_data('short.fa')
@@ -881,8 +887,10 @@ def test_do_plot_comparison_4_output_dir(c):
     assert os.path.exists(os.path.join(output_dir, "cmp.matrix.png"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_5_force(c):
+def test_do_plot_comparison_5_force(runtmp):
+    # test -f to force display of something that's not a distance matrix
+    c = runtmp
+
     D = numpy.zeros([2, 2])
     D[0, 0] = 5
     with open(c.output('cmp'), 'wb') as fp:
@@ -896,8 +904,10 @@ def test_do_plot_comparison_5_force(c):
     assert c.last_result.status == 0
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_4_fail_not_distance(c):
+def test_do_plot_comparison_4_fail_not_distance(runtmp):
+    # plot should fail when not a distance matrix
+    c = runtmp
+
     D = numpy.zeros([2, 2])
     D[0, 0] = 5
     with open(c.output('cmp'), 'wb') as fp:
@@ -914,31 +924,33 @@ def test_do_plot_comparison_4_fail_not_distance(c):
 
 
 def test_plot_override_labeltext(runtmp):
-        testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
-        testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
-        testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
-        testdata4 = utils.get_test_data('genome-s10+s11.sig')
+    # test overriding labeltext
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
 
-        runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
[…]
