### 2023-09-29 958e419
add a bunch of tests
