### 2023-09-29 a64edbf
write a few basic tests
