### 2023-09-30 fb45b6c
add information about save formats and memory usage
