### 2023-09-30 69dfbc0
add section on choosing a reference for read mapping
