+++ doc/conf.py
@@ -59,8 +59,8 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'sourmash'
-copyright = '2016-2020, C. Titus Brown and Luiz Irber'
-author = 'C. Titus Brown and Luiz Irber'
+copyright = '2016-2023, C. Titus Brown, Luiz Irber, and N. Tessa Pierce-Ward'
+author = 'C. Titus Brown, Luiz Irber, and N. Tessa Pierce-Ward'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
