### 2023-09-30 3a13ab2
Build manifest from paths in parallel
