### 2023-10-03 a4c1244
clean up missing refs
