### 2023-10-03 922ee39
Bump bytecount from 0.6.3 to 0.6.4
Bumps [bytecount](https://github.com/llogiq/bytecount) from 0.6.3 to 0.6.4.
- [Commits](https://github.com/llogiq/bytecount/commits)

---
updated-dependencies:
- dependency-name: bytecount
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
