+++ src/core/Cargo.toml
@@ -56,7 +56,7 @@ ouroboros = "0.18.0"
 assert_matches = "1.3.0"
 criterion = "0.5.1"
 needletail = { version = "0.5.1", default-features = false }
-proptest = { version = "1.2.0", default-features = false, features = ["std"]}
+proptest = { version = "1.3.1", default-features = false, features = ["std"]}
 rand = "0.8.2"
 tempfile = "3.7.1"
 
