### 2023-10-03 2245bad
Bump pypa/cibuildwheel from 2.16.0 to 2.16.1 (#2797)
Bumps [pypa/cibuildwheel](https://github.com/pypa/cibuildwheel) from
2.16.0 to 2.16.1.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/pypa/cibuildwheel/releases">pypa/cibuildwheel's
releases</a>.</em></p>
<blockquote>
<h2>v2.16.1</h2>
<ul>
<li>🛠 Updates the prerelease CPython 3.12 version to 3.12.0rc3 (<a
href="https://redirect.github.com/pypa/cibuildwheel/issues/1625">#1625</a>)</li>
<li>🛠 Only calls <code>linux32</code> in containers when necessary (<a
href="https://redirect.github.com/pypa/cibui
[…]
