+++ src/core/Cargo.toml
@@ -26,7 +26,7 @@ maturin = []
 [dependencies]
 az = "1.0.0"
 bytecount = "0.6.4"
-byteorder = "1.4.3"
+byteorder = "1.5.0"
 cfg-if = "1.0"
 counter = "0.5.7"
 finch = { version = "0.6.0", optional = true }
