+++ src/core/Cargo.toml
@@ -49,7 +49,7 @@ typed-builder = "0.14.0"
 twox-hash = "1.6.0"
 vec-collections = "0.3.4"
 piz = "0.5.0"
-memmap2 = "0.8.0"
+memmap2 = "0.9.0"
 ouroboros = "0.18.0"
 
 [dev-dependencies]
