### 2023-10-09 24bf474
Bump pypa/cibuildwheel from 2.16.1 to 2.16.2 (#2808)
Bumps [pypa/cibuildwheel](https://github.com/pypa/cibuildwheel) from
2.16.1 to 2.16.2.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/pypa/cibuildwheel/releases">pypa/cibuildwheel's
releases</a>.</em></p>
<blockquote>
<h2>v2.16.2</h2>
<ul>
<li>🛠 Updates CPython 3.12 version to 3.12.0, final release (<a
href="https://redirect.github.com/pypa/cibuildwheel/issues/1635">#1635</a>)</li>
<li>✨ Adds a debug option <a
href="https://cibuildwheel.readthedocs.io/en/stable/options/#cibw_debug_keep_container"><
[…]
