+++ doc/index.md
@@ -1,191 +1,114 @@
 # Welcome to sourmash!
 
-sourmash is a command-line tool and Python library for computing
-[hash sketches](https://en.wikipedia.org/wiki/MinHash) from DNA
-sequences, comparing them to each other, and plotting the results.
-This allows you to estimate sequence similarity between even very
-large data sets quickly and accurately.
-
-sourmash can be used to quickly search large databases of genomes
-for matches to query genomes and metagenomes; see [our list of
-available databases](databases.md).
-
-sourmash also includes k-mer based taxonomic exploration and
-classification routines for genome and metagenome analysis. These
-routines can use the NCBI and GTDB taxonomies but do not depend on them
-specifically.
-
-We have [several tutorials](tutorials.md) available! Start with
-[Making signatures, comparing, and searching](tutorial-basic.md).
-
-The paper [Large-scale sequence comparisons with sourmash (Pierce et al., 2019)](https://f1000research.com/articles/8-1006)
-gives an overview of how sourmash works and what its major use cases are.
-Please also see the `mash` [software](http://mash.readthedocs.io/en/latest/) and
-[paper (Ondov et al., 2016)](http://dx.doi.org/10.1186/s13059-016-0997-x) for
-background information on how and why MinHash works.
-
-**Questions? Thoughts?** Ask us on the [sourmash issue tracker](https://github.com/sourmash-bio/sourmash/issues/)!
-
-**Want to migrate to sourmash v4?** sourmash v4 is now available, and
-has a number of incompatibilites with v2 and v3. Please see 
-[our migration guide](support.md#migrating-from-sourmash-v3x-to-sourmash-v4x)!
-
-----
-
-To use sourmash, you must be comfortable with the UNIX command line;
-programmers may find the [Python library and API](api.md) useful as well.
-
-If you use sourmash, please cite us!
-
-> Brown and Irber (2016),
-> **sourmash: a library for MinHash sketching of DNA**.
-> Journal of Open Source Software, 1(5), 27, [doi:10.21105/joss.00027](https://joss.theoj.org/papers/3d793c6e7db683bee7c03377a4a7f3c9)
-
-## sourmash in brief
-
-sourmash uses MinHash-style sketching to create "signatures", compressed
-representations of DNA/RNA sequence.  These signatures can then
-be stored, searched, explored, and taxonomically annotated.
-
-* `sourmash` provides command line utilities for creating, comparing,
-  and searching signatures, as well as plotting and clustering
-  signatures by similarity (see [the command-line docs](command-line.md)).
-
-* `sourmash` can **search very large collections of signatures** to find matches
-  to a query.
-
-* `sourmash` can also **identify parts of metagenomes that match known genomes**,
-  and can **taxonomically classify genomes and metagenomes** against databases
-  of known species.
-
-* `sourmash` can be used to **search databases of public sequences**
-  (e.g. all of GenBank) and can also be used to create and search databases
-  of **private sequencing data**.
-
-* `sourmash` supports saving, loading, and communication of signatures
-  via [JSON](http://www.json.org/), a ~human-readable and editable format.
-
-* `sourmash` also has a simple Python API for interacting with signatures,
-  including support for online updating and querying of signatures
-  (see [the API docs](api.md)).
-
-* `sourmash` relies on an underlying Rust core for performance.
-
-* `sourmash` is developed [on GitHub](https://github.com/sourmash-bio/sourmash)
-  and is **freely and openly available** under the BSD 3-clause license.
-  Please see [the README](https://github.com/sourmash-bio/sourmash/blob/latest/README.md)
-  for more information on development, support, and contributing.
-
-You can take a look at sourmash analyses on real data
[…]
