+++ doc/new.md
@@ -1,114 +1,0 @@
-# Welcome to sourmash!
-
-```{contents} Contents
-:depth: 3
-```
-
-sourmash is a command-line tool and Python/Rust library for
-**metagenome analysis** and **genome comparison** with k-mers.  It
-supports the compositional analysis of metagenomes, rapid search of
-large sequence databases, and flexible taxonomic analysis with both
-NCBI and GTDB taxonomies. sourmash works well with sequences 30kb or
-larger, including bacterial and viral genomes.
-
-You might try sourmash if you want to -
-
-* identify which reference genomes to map your metagenomic reads to
-* search all Genbank microbial genomes with a sequence query
-* cluster many genomes by similarity
-* taxonomically classify genomes or metagenomes against NCBI and/or GTDB;
-* search thousands of metagenomes with a query genome or sequence
-
-Our **vision**: sourmash strives to support biologists in analyzing
-modern sequencing data at high resolution and with full context,
-including all public reference genomes and metagenomes.
-
-## How does sourmash work?
-
-Underneath, sourmash uses [FracMinHash sketches](https://www.biorxiv.org/content/10.1101/2022.01.11.475838) for fast and
-lightweight sequence comparison; FracMinHash builds on
-[MinHash sketching](https://en.wikipedia.org/wiki/MinHash) to support both Jaccard similarity
-_and_ containment analyses with k-mers.  This significantly expands
-the range of operations that can be done quickly and in low
-memory. sourmash also implements a number of new and powerful analysis
-techniques, including minimum metagenome covers and alignment-free ANI
-estimation.
-
-sourmash is inspired by [mash](https://mash.readthedocs.io), and
-supports most mash analyses. sourmash also implements an expanded set
-of functionality for metagenome and taxonomic analysis.
-
-sourmash development was initiated with a grant from the Moore
-Foundation under the Data Driven Discovery program, and has been
-supported by further funding from the NIH and NSF. Please see
-[funding acknowledgements](funding.md) for details!
-
-## Mission statement
-
-The project mission is to provide practical tools and approaches for
-analyzing extremely large sequencing data sets, with an emphasis on
-high resolution results. We design around the following principles:
-
-* genomic and metagenomic analyses should be able to make use of all
-  available reference genomes.
-* metagenomic analyses should support assembly independent approaches,
-  to avoid biases stemming from low coverage or high strain
-  variability.
-* private and public databases should be equally well supported.
-* a variety of data structures and algorithms are necessary to support
-  a wide set of use cases, including efficient command-line analysis,
-  real-time queries, and massive-scale batch analyses.
-* our tools should be well behaved members of the bioinformatics
-  analysis tool ecosystem, and use common installation approaches,
-  standard formats, and semantic versioning.
-* our tools should be robustly tested, well documented, and supported.
-* we discuss scientific and computational tradeoffs and make specific
-  recommendations where possible, admitting uncertainty as needed.
-
-## Using sourmash
-
-### Tutorials and examples
-
-These tutorials are command line tutorials that should work on Mac OS
-X and Linux. They require about 5 GB of disk space and 5 GB of RAM.
-
-* [The first sourmash tutorial - making signatures, comparing, and searching](tutorial-basic.md)
-
-* [Using sourmash LCA to do taxonomic classification](tutorials-lca.md)
-
-* [Analyzing the genomic and taxonomic composition of an environmental genome using GTDB and sample-specific MAGs with sourmash](tutorial-lemonade.md)
-
-* [Some sourmash command line examples!](sourmash-examples.ipynb)
-
-### How-To Guides
-
-* Installing sourmash
-
[…]
