### 2023-10-13 6026ecb
Apply suggestions from code review
thanks!

Co-authored-by: ccbaumler <63077899+ccbaumler@users.noreply.github.com>
