+++ doc/sidebar.md
@@ -1,0 +1,46 @@
+## Tutorials and examples
+
+These tutorials are command line tutorials that should work on Mac OS
+X and Linux. They require about 5 GB of disk space and 5 GB of RAM.
+
+* [The first sourmash tutorial - making signatures, comparing, and searching](tutorial-basic.md)
+
+* [Using sourmash LCA to do taxonomic classification](tutorials-lca.md)
+
+* [Analyzing the genomic and taxonomic composition of an environmental genome using GTDB and sample-specific MAGs with sourmash](tutorial-lemonade.md)
+
+* [Some sourmash command line examples!](sourmash-examples.ipynb)
+
+## How-To Guides
+
+* Installing sourmash
+
+* [Classifying genome sketches](classifying-signatures.md)
+
+* [Working with private collections of genome sketches](sourmash-collections.ipynb)
+
+* [Using the `LCA_Database` API](using-LCA-database-API.ipynb)
+
+* [Building plots from `sourmash compare` output](plotting-compare.ipynb).
+
+* [A short guide to using sourmash output with R](other-languages.md).
+
+## How sourmash works under the hood
+
+* [An introduction to k-mers for genome comparison and analysis](kmers-and-minhash.ipynb)
+* [Support, versioning, and migration between versions](support.md)
+
+## Reference material
+
+* [Full table of contents for all docs](toc.md)
+* [UNIX command-line documentation](command-line.md)
+* [Genbank and GTDB databases and taxonomy files](databases.md)
+* [Python examples using the API](api-example.md)
+* [Publications about sourmash](publications.md)
+* [A guide to the internals of sourmash](sourmash-internals.md)
+* [Funding acknowledgements](funding.md)
+
+## Developing and extending sourmash
+
+* [Releasing a new version of sourmash](release.md)
+
