### 2023-10-14 6a5bb1a
rename internals page
