+++ doc/index.md
@@ -106,7 +106,7 @@ X and Linux. They require about 5 GB of disk space and 5 GB of RAM.
 * [Genbank and GTDB databases and taxonomy files](databases.md)
 * [Python examples using the API](api-example.md)
 * [Publications about sourmash](publications.md)
-* [A guide to the internals of sourmash](sourmash-internals.md)
+* [A guide to the internal design and structure of sourmash](sourmash-internals.md)
 * [Funding acknowledgements](funding.md)
 
 ## Developing and extending sourmash
