### 2023-10-15 9ff523e
MRG: make `sourmash plot` labels/indices arguments make sense (#2790)
This PR rationalizes `sig plot` arguments for `--labels` (show names)
and `--indices` (show numbers), and adds `--no-labels` and
`--no-indices`, as follows. See
https://github.com/sourmash-bio/sourmash/issues/2667 for motivating bug.

1. `sourmash plot compare-demo` - labels on dendrogram, labels on matrix
✅ (FIXED)

2. `sourmash plot compare-demo --labels` - labels on both dendrogram and
matrix ✅

3. `sourmash plot compare-demo --indices` - indices on both dendrogram
and matrix ✅

4. `sourmash plot compare-demo --labels --i
[…]
