### 2023-10-15 32ecc0d
clean up dev instructions a bit
