### 2023-10-15 0db31c8
MRG: revise documentation structure; add internals page. (#2184)
This PR rearranges the docs to according to the https://diataxis.fr/
structure, per https://github.com/sourmash-bio/sourmash/issues/2054.

New pages:
* [A heavily revised index
page](https://sourmash--2184.org.readthedocs.build/en/2184/index.html)
* [A guide to the internals of
sourmash](https://sourmash--2184.org.readthedocs.build/en/2184/sourmash-internals.html)
* [Frequently asked
questions](https://sourmash--2184.org.readthedocs.build/en/2184/faq.html)
*
[Publications](https://sourmash--2184.org.readthedocs.build/en/2184/publ
[…]
