+++ doc/index.md
@@ -1,191 +1,133 @@
 # Welcome to sourmash!
 
-sourmash is a command-line tool and Python library for computing
-[hash sketches](https://en.wikipedia.org/wiki/MinHash) from DNA
-sequences, comparing them to each other, and plotting the results.
-This allows you to estimate sequence similarity between even very
-large data sets quickly and accurately.
+sourmash is a command-line tool and Python/Rust library for
+**metagenome analysis** and **genome comparison** using k-mers.  It
+supports the compositional analysis of metagenomes, rapid search of
+large sequence databases, and flexible taxonomic profiling with both
+NCBI and GTDB taxonomies
+([see our prepared databases for more information](databases.md)). sourmash
+works well with sequences 30kb or larger, including bacterial and
+viral genomes.
 
-sourmash can be used to quickly search large databases of genomes
-for matches to query genomes and metagenomes; see [our list of
-available databases](databases.md).
+You might try sourmash if you want to -
 
-sourmash also includes k-mer based taxonomic exploration and
-classification routines for genome and metagenome analysis. These
-routines can use the NCBI and GTDB taxonomies but do not depend on them
-specifically.
+* identify which reference genomes to map your metagenomic reads to
+* search all Genbank microbial genomes with a sequence query
+* cluster many genomes by similarity
+* taxonomically classify genomes or metagenomes against NCBI and/or GTDB
+* search thousands of metagenomes with a query genome or sequence
 
-We have [several tutorials](tutorials.md) available! Start with
-[Making signatures, comparing, and searching](tutorial-basic.md).
+Our **vision**: sourmash strives to support biologists in analyzing
+modern sequencing data at high resolution and with full context,
+including all public reference genomes and metagenomes.
 
-The paper [Large-scale sequence comparisons with sourmash (Pierce et al., 2019)](https://f1000research.com/articles/8-1006)
-gives an overview of how sourmash works and what its major use cases are.
-Please also see the `mash` [software](http://mash.readthedocs.io/en/latest/) and
-[paper (Ondov et al., 2016)](http://dx.doi.org/10.1186/s13059-016-0997-x) for
-background information on how and why MinHash works.
+## Mission statement
 
-**Questions? Thoughts?** Ask us on the [sourmash issue tracker](https://github.com/sourmash-bio/sourmash/issues/)!
+This project's mission is to provide practical tools and approaches for
+analyzing extremely large sequencing data sets, with an emphasis on
+high resolution results. Our designs follow these guiding principles:
 
-**Want to migrate to sourmash v4?** sourmash v4 is now available, and
-has a number of incompatibilites with v2 and v3. Please see 
-[our migration guide](support.md#migrating-from-sourmash-v3x-to-sourmash-v4x)!
+* genomic and metagenomic analyses should be able to make use of all
+  available reference genomes.
+* metagenomic analyses should support assembly independent approaches,
+  to avoid biases stemming from low coverage or high strain
+  variability.
+* private and public databases should be equally well supported.
+* a variety of data structures and algorithms are necessary to support
+  a wide set of use cases, including efficient command-line analysis,
+  real-time queries, and massive-scale batch analyses.
+* our tools should be well behaved members of the bioinformatics
+  analysis tool ecosystem, and use common installation approaches,
+  standard formats, and semantic versioning.
+* our tools should be robustly tested, well documented, and supported.
+* we discuss scientific and computational tradeoffs and make specific
+  recommendations where possible, admitting uncertainty as needed.
 
[…]
