### 2023-10-16 009e462
MRG: fix many minor documentation nits (#2814)
Follow-up to #2184 with minor corrections.
