+++ doc/command-line.md
@@ -5,7 +5,7 @@
 ```
 
 From the command line, sourmash can be used to create
-[MinHash sketches][0] from DNA and protein sequences, compare them to
+[FracMinHash sketches][0] from DNA and protein sequences, compare them to
 each other, and plot the results; these sketches are saved into
 "signature files".  These signatures allow you to estimate sequence
 similarity and containment quickly and accurately in large
@@ -45,7 +45,10 @@ Next, compare all the signatures to each other:
 sourmash compare *.sig -o cmp.dist
 ```
 
-Finally, plot a dendrogram: ``` sourmash plot cmp.dist --labels ```
+Finally, plot a dendrogram:
+```
+sourmash plot cmp.dist --labels
+```
 This will output three files, `cmp.dist.dendro.png`,
 `cmp.dist.matrix.png`, and `cmp.dist.hist.png`, containing a
 clustering & dendrogram of the sequences, a similarity matrix and
@@ -562,15 +565,16 @@ As with all reference-based analysis, results can be affected by the
  and redundancy of reference databases.
 
 For more details on how `gather` works and can be used to classify
- signatures, see [classifying-signatures](classifying-signatures.md).
+signatures, see <project:classifying-signatures.md>
 
 ### `sourmash tax metagenome` - summarize metagenome content from `gather` results
 
 `sourmash tax metagenome` summarizes gather results for each query metagenome by
- taxonomic lineage.
+taxonomic lineage.
 
-example command to summarize a single `gather csv`, where the query was gathered
- against `gtdb-rs202` representative species database:
+Here is an example command to summarize a single `gather csv`, where
+ the query was gathered against `gtdb-rs202` representative species
+database:
 
 ```
 sourmash tax metagenome
@@ -589,10 +593,10 @@ The possible output formats are:
 #### `csv_summary` output format
 
 `csv_summary` is the default output format. This outputs a `csv` with lineage
- summarization for each taxonomic rank. This output currently consists of six
- columns, `query_name,rank,fraction,lineage,query_md5,query_filename`, where
- `fraction` is the  fraction of the query matched to the reported rank and
- lineage.
+summarization for each taxonomic rank. This output currently consists of six
+columns, `query_name,rank,fraction,lineage,query_md5,query_filename`, where
+`fraction` is the  fraction of the query matched to the reported rank and
+lineage.
 
 example `csv_summary` output from the command above:
 
@@ -611,7 +615,7 @@ o__Bacteroidales;f__Bacteroidaceae;g__Prevotella;s__Prevotella copri
 HSMA33MX,species,0.016,d__Bacteria;p__Bacteroidota;c__Bacteroidia;
 o__Bacteroidales;f__Bacteroidaceae;g__Phocaeicola;s__Phocaeicola vulgatus
 ```
-> The `query_md5` and `query_filename` columns are omitted here for brevity.
+The `query_md5` and `query_filename` columns are omitted here for brevity.
 
 #### `krona` output format
 
@@ -1005,14 +1009,14 @@ commands.
 
 All `sourmash tax` commands must be given one or more taxonomy files as
 parameters to the `--taxonomy` argument. These files can be either CSV
-files or (as of sourmash 4.2.1) sqlite3 databases. sqlite3 databases
+files or (as of sourmash 4.2.1) SQLite databases. SQLite databases
 are much faster for large taxonomies, while CSV files are easier to view
 and modify using spreadsheet software.
 
 `sourmash tax prepare` is a utility function that can ingest and validate
-multiple CSV files or sqlite3 databases, and output a CSV file or a sqlite3
+multiple CSV files or SQLite databases, and output a CSV file or a SQLite
[…]
