### 2023-10-16 fe7ca34
Bump typed-builder from 0.14.0 to 0.17.0
Bumps [typed-builder](https://github.com/idanarye/rust-typed-builder) from 0.14.0 to 0.17.0.
- [Changelog](https://github.com/idanarye/rust-typed-builder/blob/master/CHANGELOG.md)
- [Commits](https://github.com/idanarye/rust-typed-builder/commits)

---
updated-dependencies:
- dependency-name: typed-builder
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
