+++ src/core/Cargo.toml
@@ -25,7 +25,7 @@ maturin = []
 
 [dependencies]
 az = "1.0.0"
-bytecount = "0.6.4"
+bytecount = "0.6.5"
 byteorder = "1.5.0"
 cfg-if = "1.0"
 counter = "0.5.7"
