### 2023-10-29 e18489c
add explicit install tutorial
