### 2023-10-29 adce355
add bioconda channel
