### 2023-10-30 51d8481
Bump bytecount from 0.6.5 to 0.6.7 (#2830)
Bumps [bytecount](https://github.com/llogiq/bytecount) from 0.6.5 to
0.6.7.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/llogiq/bytecount/commits">compare view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibility
score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=bytecount&package-manager=cargo&previous-version=0.6.5&new-version=0.6.7)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-updates#about-compatibility-s
[…]
