### 2023-10-30 26628a5
Bump tempfile from 3.8.0 to 3.8.1 (#2828)
[//]: # (dependabot-start)
⚠️  **Dependabot is rebasing this PR** ⚠️ 

Rebasing might not happen immediately, so don't worry if this takes some
time.

Note: if you make any changes to this PR yourself, they will take
precedence over the rebase.

---

[//]: # (dependabot-end)

Bumps [tempfile](https://github.com/Stebalien/tempfile) from 3.8.0 to
3.8.1.
<details>
<summary>Changelog</summary>
<p><em>Sourced from <a
href="https://github.com/Stebalien/tempfile/blob/master/CHANGELOG.md">tempfile's
changelog</a>.</em></p>
<blockquote>
<h2>3.8.1</h2>
<ul>
<li>
[…]
