+++ src/core/Cargo.toml
@@ -58,7 +58,7 @@ criterion = "0.5.1"
 needletail = { version = "0.5.1", default-features = false }
 proptest = { version = "1.3.1", default-features = false, features = ["std"]}
 rand = "0.8.2"
-tempfile = "3.7.1"
+tempfile = "3.8.1"
 
 [[bench]]
 name = "compute"
