### 2023-10-31 7e8198b
MRG: add explicit install tutorial (#2826)
Fixes https://github.com/sourmash-bio/sourmash/issues/2813

See new index and sidebar at [the docs build for this
PR](https://sourmash--2826.org.readthedocs.build/en/2826/index.html)
