### 2023-11-13 215fc04
Bump wasm-bindgen-test from 0.3.37 to 0.3.38
Bumps [wasm-bindgen-test](https://github.com/rustwasm/wasm-bindgen) from 0.3.37 to 0.3.38.
- [Release notes](https://github.com/rustwasm/wasm-bindgen/releases)
- [Changelog](https://github.com/rustwasm/wasm-bindgen/blob/main/CHANGELOG.md)
- [Commits](https://github.com/rustwasm/wasm-bindgen/commits)

---
updated-dependencies:
- dependency-name: wasm-bindgen-test
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
