### 2023-11-15 #129 @snacktavish
## Editor comments
:wave: Hi @LilyAnderssonLee and @elais! Thank you for volunteering to review
for pyOpenSci!

- [ ] @LilyAnderssonLee  survey completed.
- [ ] @elais survey completed.

Please get in touch with any questions or concerns! Your review is due: Friday, December 8
