### 2023-11-22 f07bc34
Use byteorder 1.4.3 for pyo3_branchwater compat
