### 2023-11-22 be6b038
Add a test for innerstorage save_sig
