### 2023-11-23 d2078c9
Drop 3.8 and 3.9, default to 3.10 for most CI tasks
