+++ tox.ini
@@ -36,6 +36,7 @@ pass_env =
     PIP_CACHE_DIR
     CI
     PYTHONDEVMODE
+    PYTHONTRACEMALLOC
     LIBCLANG_PATH
     BINDGEN_EXTRA_CLANG_ARGS
 deps =
