### 2023-11-25 32d46dd
Drop 3.8 and 3.9, default to 3.10 for most CI tasks (#2848)
[Numpy NEP 29](https://numpy.org/neps/nep-0029-deprecation_policy.html)
is now [Scientific Python SPEC
0](https://scientific-python.org/specs/spec-0000/), and according to
that we could drop 3.9 already. We still had many places mentioning 3.8
(and even 3.7!), so I went through all mentions to
`3.7`/`37`/`3.8`/`38`/`3.9`/`39` and removed them, updating to `3.10`
(our new lowest supported Python version) as appropriate.
