+++ tox.ini
@@ -1,12 +1,11 @@
 [tox]
 env_list =
-  py310,
   py311,
-  py39,
+  py312,
+  py310,
   coverage,
   docs,
   package_description
-  py38,
   fix_lint,
   hypothesis,
   khmer,
@@ -67,7 +66,7 @@ commands = pytest \
            {posargs:.}
 
 [testenv:khmer]
-basepython = python3.8
+basepython = python3.10
 deps =
   khmer
 commands = pytest \
@@ -79,7 +78,7 @@ commands = pytest \
            {posargs:.}
 
 [testenv:khmer_master]
-basepython = python3.8
+basepython = python3.10
 deps =
   -e git+https://github.com/dib-lab/khmer.git\#egg=khmer
 commands = pytest \
@@ -113,7 +112,7 @@ commands = sphinx-build -d "{toxworkdir}/docs_doctree" doc "{toxworkdir}/docs_ou
 
 [testenv:package_description]
 description = check that the long description is valid
-basepython = python3.8
+basepython = python3.10
 deps = twine >= 1.12.1
        # TODO installing readme-renderer[md] should not be necessary
        readme-renderer[md] >= 24.0
@@ -126,7 +125,7 @@ commands = pip wheel -w {envtmpdir}/build --no-deps .
 
 [testenv:mypy]
 description = run mypy checker
-basepython = python3.8
+basepython = python3.10
 pass_env = {[testenv]pass_env}
           # without PROGRAMDATA cloning using git for Windows will fail with an `error setting certificate verify locations` error
           PROGRAMDATA
@@ -135,7 +134,7 @@ commands = mypy src/sourmash
 
 [testenv:fix_lint]
 description = format the code base to adhere to our styles, and complain about what we cannot do automatically
-basepython = python3.8
+basepython = python3.10
 pass_env = {[testenv]pass_env}
           # without PROGRAMDATA cloning using git for Windows will fail with an `error setting certificate verify locations` error
           PROGRAMDATA
@@ -161,7 +160,7 @@ commands = coverage combine
            coverage xml -i -o {toxworkdir}/coverage.xml
            coverage html -i -d {toxworkdir}/htmlcov
            diff-cover --compare-branch {env:DIFF_AGAINST:origin/latest} {toxworkdir}/coverage.xml
-depends = py39, py38, py37, pypy3
+depends = py312, py311, py310, pypy3
 parallel_show_output = True
 
 [testenv:X]
@@ -198,8 +197,7 @@ source = src/sourmash/
 python =
   3.10: py310, docs, package_description, coverage
   3.11: py311, coverage
-  3.9: py39, coverage
-  3.8: py38, coverage
+  3.12: py312, coverage
 
 [flake8]
 max-complexity = 22
