+++ tests/test_tax.py
@@ -6,6 +6,7 @@ import csv
 import pytest
 import gzip
 from collections import Counter
+from pathlib import Path
 
 import sourmash
 import sourmash_tst_utils as utils
@@ -118,7 +119,7 @@ def test_metagenome_summary_csv_out(runtmp):
     assert runtmp.last_result.status == 0
     assert os.path.exists(csvout)
 
-    sum_gather_results = [x.rstrip() for x in open(csvout)]
+    sum_gather_results = [x.rstrip() for x in Path(csvout).read_text().splitlines()]
     assert f"saving 'csv_summary' output to '{csvout}'" in runtmp.last_result.err
     assert 'query_name,rank,fraction,lineage,query_md5,query_filename,f_weighted_at_rank,bp_match_at_rank' in sum_gather_results[0]
     assert 'test1,superkingdom,0.2042281611487834,d__Bacteria,md5,test1.sig,0.13080306238801107,1024000' in  sum_gather_results[1]
@@ -160,7 +161,7 @@ def test_metagenome_summary_csv_out_empty_gather_force(runtmp):
     print("g_csv: ", gather_empty)
 
     runtmp.run_sourmash('tax', 'metagenome', '--gather-csv', g_csv, '-g', gather_empty, '--taxonomy-csv', tax, '-o', csv_base, '--output-dir', outdir, '-f')
-    sum_gather_results = [x.rstrip() for x in open(csvout)]
+    sum_gather_results = [x.rstrip() for x in Path(csvout).read_text().splitlines()]
     assert f"saving 'csv_summary' output to '{csvout}'" in runtmp.last_result.err
     assert 'query_name,rank,fraction,lineage,query_md5,query_filename,f_weighted_at_rank,bp_match_at_rank' in sum_gather_results[0]
     assert 'test1,superkingdom,0.2042281611487834,d__Bacteria,md5,test1.sig,0.13080306238801107,1024000' in  sum_gather_results[1]
@@ -184,7 +185,7 @@ def test_metagenome_kreport_out(runtmp):
     assert runtmp.last_result.status == 0
     assert os.path.exists(csvout)
 
-    kreport_results = [x.rstrip().split('\t') for x in open(csvout)]
+    kreport_results = [x.rstrip().split('\t') for x in Path(csvout).read_text().splitlines()]
     assert f"saving 'kreport' output to '{csvout}'" in runtmp.last_result.err
     print(kreport_results)
     assert ['13.08', '1605999', '0', 'D', '', 'd__Bacteria'] == kreport_results[0]
@@ -223,7 +224,7 @@ def test_metagenome_kreport_ncbi_taxid_out(runtmp):
     assert runtmp.last_result.status == 0
     assert os.path.exists(csvout)
 
-    kreport_results = [x.rstrip().split('\t') for x in open(csvout)]
+    kreport_results = [x.rstrip().split('\t') for x in Path(csvout).read_text().splitlines()]
     assert f"saving 'kreport' output to '{csvout}'" in runtmp.last_result.err
     print(kreport_results)
     assert ['13.08', '1605999', '0', 'D', '2', 'Bacteria'] == kreport_results[0]
@@ -263,7 +264,7 @@ def test_metagenome_kreport_out_lemonade(runtmp):
     assert runtmp.last_result.status == 0
     assert os.path.exists(csvout)
 
-    kreport_results = [x.rstrip().split('\t') for x in open(csvout)]
+    kreport_results = [x.rstrip().split('\t') for x in Path(csvout).read_text().splitlines()]
     assert f"saving 'kreport' output to '{csvout}'" in runtmp.last_result.err
     print(kreport_results)
[…]
