+++ tox.ini
@@ -35,6 +35,7 @@ pass_env =
     PIP_CACHE_DIR
     CI
     PYTHONDEVMODE
+    PYTHONTRACEMALLOC
     LIBCLANG_PATH
     BINDGEN_EXTRA_CLANG_ARGS
 deps =

[…]
