### 2023-11-27 f332b90
Bump conda-incubator/setup-miniconda from 2.2.0 to 3.0.0
Bumps [conda-incubator/setup-miniconda](https://github.com/conda-incubator/setup-miniconda) from 2.2.0 to 3.0.0.
- [Release notes](https://github.com/conda-incubator/setup-miniconda/releases)
- [Changelog](https://github.com/conda-incubator/setup-miniconda/blob/main/CHANGELOG.md)
- [Commits](https://github.com/conda-incubator/setup-miniconda/compare/3b0f2504dd76ef23b6d31f291f4913fb60ab5ff3...2defc80cc6f4028b1780c50faf08dd505d698976)

---
updated-dependencies:
- dependency-name: conda-incubator/setup-miniconda
  dependency-type: direct:pr
[…]
