+++ .github/workflows/dev_envs.yml
@@ -42,7 +42,7 @@ jobs:
           ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles('environment.yml') }}
 
     - name: setup conda
-      uses: conda-incubator/setup-miniconda@3b0f2504dd76ef23b6d31f291f4913fb60ab5ff3
+      uses: conda-incubator/setup-miniconda@2defc80cc6f4028b1780c50faf08dd505d698976
       with:
         auto-update-conda: true
         python-version: "3.10"
