### 2023-11-28 d20ed2e
Bump wasm-bindgen from 0.2.88 to 0.2.89 (#2850)
Bumps [wasm-bindgen](https://github.com/rustwasm/wasm-bindgen) from
0.2.88 to 0.2.89.
<details>
<summary>Changelog</summary>
<p><em>Sourced from <a
href="https://github.com/rustwasm/wasm-bindgen/blob/main/CHANGELOG.md">wasm-bindgen's
changelog</a>.</em></p>
<blockquote>
<h1><code>wasm-bindgen</code> Change Log</h1>
<hr />
<h2><a
href="https://github.com/rustwasm/wasm-bindgen/compare/0.2.88...main">Unreleased</a></h2>
<h3>Added</h3>
<ul>
<li>
<p>Add additional constructor to <code>DataView</code> for
<code>SharedArrayBuffer</code>.
<a
href="https:
[…]
