### 2023-11-28 4cb056f
add memory usage vs ksize to FAQ
