### 2023-11-28 f3f1b47
change 'sig kmers' to match documented behavior; add tests
