### 2023-11-28 6dcc515
print out problem in assert
