+++ src/core/tests/minhash.rs
@@ -18,7 +18,7 @@ const EPSILON: f64 = 0.01;
 
 #[test]
 fn throws_error() {
-    let mut mh = KmerMinHash::new(0, 4, HashFunctions::murmur64_DNA, 42, false, 1);
+    let mut mh = KmerMinHash::new(0, 4, HashFunctions::Murmur64Dna, 42, false, 1);
 
     assert!(
         mh.add_sequence(b"ATGR", false).is_err(),
@@ -28,8 +28,8 @@ fn throws_error() {
 
 #[test]
 fn merge() {
-    let mut a = KmerMinHash::new(0, 10, HashFunctions::murmur64_DNA, 42, false, 20);
-    let mut b = KmerMinHash::new(0, 10, HashFunctions::murmur64_DNA, 42, false, 20);
+    let mut a = KmerMinHash::new(0, 10, HashFunctions::Murmur64Dna, 42, false, 20);
+    let mut b = KmerMinHash::new(0, 10, HashFunctions::Murmur64Dna, 42, false, 20);
 
     a.add_sequence(b"TGCCGCCCAGCA", false).unwrap();
     b.add_sequence(b"TGCCGCCCAGCA", false).unwrap();
@@ -55,20 +55,20 @@ fn merge() {
 
 #[test]
 fn invalid_dna() {
-    let mut a = KmerMinHash::new(0, 3, HashFunctions::murmur64_DNA, 42, false, 20);
+    let mut a = KmerMinHash::new(0, 3, HashFunctions::Murmur64Dna, 42, false, 20);
 
     a.add_sequence(b"AAANNCCCTN", true).unwrap();
     assert_eq!(a.mins().len(), 3);
 
-    let mut b = KmerMinHash::new(0, 3, HashFunctions::murmur64_DNA, 42, false, 20);
+    let mut b = KmerMinHash::new(0, 3, HashFunctions::Murmur64Dna, 42, false, 20);
     b.add_sequence(b"NAAA", true).unwrap();
     assert_eq!(b.mins().len(), 1);
 }
 
 #[test]
 fn similarity() -> Result<(), Box<dyn std::error::Error>> {
-    let mut a = KmerMinHash::new(0, 20, HashFunctions::murmur64_hp, 42, true, 5);
-    let mut b = KmerMinHash::new(0, 20, HashFunctions::murmur64_hp, 42, true, 5);
+    let mut a = KmerMinHash::new(0, 20, HashFunctions::Murmur64Hp, 42, true, 5);
+    let mut b = KmerMinHash::new(0, 20, HashFunctions::Murmur64Hp, 42, true, 5);
 
     a.add_hash(1);
     b.add_hash(1);
@@ -82,8 +82,8 @@ fn similarity() -> Result<(), Box<dyn std::error::Error>> {
 
 #[test]
 fn similarity_2() -> Result<(), Box<dyn std::error::Error>> {
-    let mut a = KmerMinHash::new(0, 5, HashFunctions::murmur64_DNA, 42, true, 5);
-    let mut b = KmerMinHash::new(0, 5, HashFunctions::murmur64_DNA, 42, true, 5);
+    let mut a = KmerMinHash::new(0, 5, HashFunctions::Murmur64Dna, 42, true, 5);
+    let mut b = KmerMinHash::new(0, 5, HashFunctions::Murmur64Dna, 42, true, 5);
 
     a.add_sequence(b"ATGGA", false)?;
     a.add_sequence(b"GGACA", false)?;
@@ -102,8 +102,8 @@ fn similarity_2() -> Result<(), Box<dyn std::error::Error>> {
 
 #[test]
 fn similarity_3() -> Result<(), Box<dyn std::error::Error>> {
[…]
