+++ src/core/src/index/revindex/mem_revindex.rs
@@ -1,0 +1,461 @@
+use std::sync::atomic::{AtomicUsize, Ordering};
+
+use camino::Utf8Path as Path;
+use camino::Utf8PathBuf as PathBuf;
+use log::{debug, info};
+
+#[cfg(feature = "parallel")]
+use rayon::prelude::*;
+
+use crate::collection::Collection;
+use crate::encodings::{Colors, Idx};
+use crate::index::linear::LinearIndex;
+use crate::index::revindex::HashToColor;
+use crate::index::{GatherResult, Index, SigCounter};
+use crate::prelude::*;
+use crate::signature::{Signature, SigsTrait};
+use crate::sketch::minhash::KmerMinHash;
+use crate::sketch::Sketch;
+use crate::Result;
+
+pub struct RevIndex {
+    linear: LinearIndex,
+    hash_to_color: HashToColor,
+    colors: Colors,
+}
+
+impl LinearIndex {
+    fn index(
+        self,
+        threshold: usize,
+        merged_query: Option<KmerMinHash>,
+        queries: Option<&[KmerMinHash]>,
+    ) -> RevIndex {
+        let processed_sigs = AtomicUsize::new(0);
+
+        #[cfg(feature = "parallel")]
+        let sig_iter = self.collection().par_iter();
+
+        #[cfg(not(feature = "parallel"))]
+        let sig_iter = self.collection().iter();
+
+        let filtered_sigs = sig_iter.enumerate().filter_map(|(dataset_id, _)| {
+            let i = processed_sigs.fetch_add(1, Ordering::SeqCst);
+            if i % 1000 == 0 {
+                info!("Processed {} reference sigs", i);
+            }
+
+            let search_sig = self
+                .collection()
+                .sig_for_dataset(dataset_id as Idx)
+                .expect("Error loading sig")
+                .into();
+
+            RevIndex::map_hashes_colors(
+                dataset_id as Idx,
+                &search_sig,
+                queries,
+                &merged_query,
+                threshold,
+                self.template(),
+            )
+        });
+
+        #[cfg(feature = "parallel")]
+        let (hash_to_color, colors) = filtered_sigs.reduce(
+            || (HashToColor::new(), Colors::default()),
+            HashToColor::reduce_hashes_colors,
+        );
+
+        #[cfg(not(feature = "parallel"))]
+        let (hash_to_color, colors) = filtered_sigs.fold(
+            (HashToColor::new(), Colors::default()),
+            HashToColor::reduce_hashes_colors,
+        );
+
+        RevIndex {
+            hash_to_color,
+            colors,
+            linear: self,
+        }
+    }
+}
+
+impl RevIndex {
+    pub fn new(
+        search_sigs: &[PathBuf],
+        selection: &Selection,
+        threshold: usize,
+        queries: Option<&[KmerMinHash]>,
+        _keep_sigs: bool,
+    ) -> Result<RevIndex> {
+        // If threshold is zero, let's merge all queries and save time later
+        let merged_query = queries.and_then(|qs| Self::merge_queries(qs, threshold));
+
+        let collection = Collection::from_paths(search_sigs)?.select(&selection)?;
+        let linear = LinearIndex::from_collection(collection.try_into()?);
+
+        Ok(linear.index(threshold, merged_query, queries))
+    }
+
+    pub fn from_zipfile<P: AsRef<Path>>(
+        zipfile: P,
+        selection: &Selection,
+        threshold: usize,
+        queries: Option<&[KmerMinHash]>,
+        _keep_sigs: bool,
+    ) -> Result<RevIndex> {
+        // If threshold is zero, let's merge all queries and save time later
+        let merged_query = queries.and_then(|qs| Self::merge_queries(qs, threshold));
+
+        let collection = Collection::from_zipfile(zipfile)?.select(&selection)?;
[…]
