### 2023-12-03 484e148
Update typed-builder and histogram, bump MSRV to 1.65
