+++ src/core/Cargo.toml
@@ -11,7 +11,7 @@ edition = "2021"
 readme = "README.md"
 autoexamples = false
 autobins = false
-rust-version = "1.64.0"
+rust-version = "1.65.0"
 
 [lib]
 name = "sourmash"
@@ -37,7 +37,7 @@ finch = { version = "0.6.0", optional = true }
 fixedbitset = "0.4.0"
 getrandom = { version = "0.2", features = ["js"] }
 getset = "0.1.1"
-histogram = "0.6.9"
+histogram = "0.8.3"
 log = "0.4.20"
 md5 = "0.7.0"
 memmap2 = "0.9.0"
@@ -56,7 +56,7 @@ serde = { version = "1.0.168", features = ["derive"] }
 serde_json = "1.0.108"
 thiserror = "1.0"
 twox-hash = "1.6.0"
-typed-builder = "0.14.0"
+typed-builder = "0.18.0"
 vec-collections = "0.4.3"
 
 [dev-dependencies]
