### 2023-12-03 ddabcad
Fix wheel building and semver checks post r0.12.0 (#2857)
`latest` checks are broken due to a misconfiguration of
`cargo-semver-checks` and trying to build wheels for platforms where
rocksdb is not being built yet (`ppc64le` and `s390x`).
