+++ pyproject.toml
@@ -164,7 +164,7 @@ known_first_party = ["sourmash"]
 
 [tool.cibuildwheel]
 build = "cp310-*"
-skip = "*-win32 *-manylinux_i686 *-musllinux_*"
+skip = "*-win32 *-manylinux_i686 *-musllinux_* *-ppc64le *-s390x"
 before-all = [
   "curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain=stable",
   "cargo update --dry-run",
