+++ src/core/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+Fixed:
+
+- CI: Fix wheel building and semver checks post r0.12.0 (#2857)
+
 ## [0.12.0] - 2023-11-26
 
 MSRV: 1.64
