+++ src/core/README.md
@@ -38,4 +38,4 @@ Development happens on github at
 
 ## Minimum supported Rust version
 
-Currently the minimum supported Rust version is 1.64.0.
+Currently the minimum supported Rust version is 1.65.0.
