### 2023-12-04 a395412
Bump histogram from 0.6.9 to 0.8.3
Bumps [histogram](https://github.com/pelikan-io/rustcommon) from 0.6.9 to 0.8.3.
- [Commits](https://github.com/pelikan-io/rustcommon/commits)

---
updated-dependencies:
- dependency-name: histogram
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
