+++ src/core/Cargo.toml
@@ -37,7 +37,7 @@ finch = { version = "0.6.0", optional = true }
 fixedbitset = "0.4.0"
 getrandom = { version = "0.2", features = ["js"] }
 getset = "0.1.1"
-histogram = "0.6.9"
+histogram = "0.8.3"
 log = "0.4.20"
 md5 = "0.7.0"
 memmap2 = "0.9.0"
