### 2023-12-05 9529c73
Deps: update typed-builder and histogram, bump MSRV to 1.65 (#2858)
Small MSRV bump to move to newer versions of `typed-builder` and
`histogram`.

Update nix flake deps.

Also disable CI jobs from #2857 until wheel building is fixed.
