+++ .github/workflows/rust.yml
@@ -250,11 +250,11 @@ jobs:
 
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: "1.64.0"
+          toolchain: "1.65.0"
           override: true
 
       - name: check if README matches MSRV defined here
-        run: grep '1.64.0' src/core/README.md
+        run: grep '1.65.0' src/core/README.md
 
       - name: Check if it builds properly
         uses: actions-rs/cargo@v1
