+++ .readthedocs.yml
@@ -8,7 +8,7 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.10"
-    rust: "1.64"
+    rust: "1.70"
   apt_packages:
     - llvm-dev
     - libclang-dev
