### 2023-12-05 00dc651
Bump proptest from 1.3.1 to 1.4.0 (#2837)
Bumps [proptest](https://github.com/proptest-rs/proptest) from 1.3.1 to
1.4.0.
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/proptest-rs/proptest/commit/63ef67c71ff050b473d2137fd3abce6115d0216b"><code>63ef67c</code></a>
[Release] 1.4.0 : adds missing changelog entries (<a
href="https://redirect.github.com/proptest-rs/proptest/issues/397">#397</a>)</li>
<li><a
href="https://github.com/proptest-rs/proptest/commit/3d40220e74cd343ecb9a0d0bf4566efb45ee58d5"><code>3d40220</code></a>
Merge pull request <a
href="https://redirect.git
[…]
