+++ src/core/Cargo.toml
@@ -62,7 +62,7 @@ vec-collections = "0.4.3"
 [dev-dependencies]
 criterion = "0.5.1"
 needletail = { version = "0.5.1", default-features = false }
-proptest = { version = "1.3.1", default-features = false, features = ["std"]}
+proptest = { version = "1.4.0", default-features = false, features = ["std"]}
 rand = "0.8.2"
 tempfile = "3.8.1"
 
