### 2023-12-07 a29e9ed
Bump byteorder from 1.4.3 to 1.5.0 (#2866)
Bumps [byteorder](https://github.com/BurntSushi/byteorder) from 1.4.3 to
1.5.0.
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/BurntSushi/byteorder/commit/ec068eefa042d494475db125c4b034bd8e9e34dd"><code>ec068ee</code></a>
1.5.0</li>
<li><a
href="https://github.com/BurntSushi/byteorder/commit/0cfb84ddc5116e49052750ff044805de0d5d17d3"><code>0cfb84d</code></a>
cargo: a few small fixups</li>
<li><a
href="https://github.com/BurntSushi/byteorder/commit/62b0e31f277e34e89fb7060f6bae5be16f05d00c"><code>62b0e31</code></a>
changelog: a
[…]
