### 2023-12-07 5659a9b
Bump web-sys from 0.3.65 to 0.3.66 (#2864)
Bumps [web-sys](https://github.com/rustwasm/wasm-bindgen) from 0.3.65 to
0.3.66.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/rustwasm/wasm-bindgen/commits">compare
view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibility
score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=web-sys&package-manager=cargo&previous-version=0.3.65&new-version=0.3.66)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-updates#about-compa
[…]
