### 2023-12-07 6ae87a7
Bump roaring from 0.10.1 to 0.10.2 (#2865)
Bumps [roaring](https://github.com/RoaringBitmap/roaring-rs) from 0.10.1
to 0.10.2.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/RoaringBitmap/roaring-rs/releases">roaring's
releases</a>.</em></p>
<blockquote>
<h2>Version 0.10.2</h2>
<h2>Introduce the <code>RoaringBitmap::remove_smallest/biggests</code>
methods</h2>
<p>Thanks to <a
href="https://github.com/shimatar0"><code>@​shimatar0</code></a>, who
added those two methods. They remove the <code>n</code> smallest or
biggest values from a <code>RoaringBitm
[…]
