+++ src/core/Cargo.toml
@@ -50,7 +50,7 @@ ouroboros = "0.18.0"
 piz = "0.5.0"
 primal-check = "0.3.1"
 rkyv = { version = "0.7.39", optional = true }
-roaring = "0.10.0"
+roaring = "0.10.2"
 rayon = { version = "1.8.0", optional = true }
 serde = { version = "1.0.168", features = ["derive"] }
 serde_json = "1.0.108"
