### 2023-12-07 e47c18c
Bump csv from 1.2.0 to 1.3.0 (#2862)
Bumps [csv](https://github.com/BurntSushi/rust-csv) from 1.2.0 to 1.3.0.
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/BurntSushi/rust-csv/commit/6374614ef8240d110626409ae882339b8145282c"><code>6374614</code></a>
1.3.0</li>
<li><a
href="https://github.com/BurntSushi/rust-csv/commit/e123861dbffd496ed621dcf9dd789150b54e9c91"><code>e123861</code></a>
deps: bump csv-core to 0.1.11</li>
<li><a
href="https://github.com/BurntSushi/rust-csv/commit/44b836cf9216a22bf150aecf7b7f0918fd96e680"><code>44b836c</code></a>
csv-core-0.1.11</li>
<li
[…]
