+++ src/core/Cargo.toml
@@ -31,7 +31,7 @@ byteorder = "1.5.0"
 camino = { version = "1.1.6", features = ["serde1"] }
 cfg-if = "1.0"
 counter = "0.5.7"
-csv = "1.1.6"
+csv = "1.3.0"
 enum_dispatch = "0.3.12"
 finch = { version = "0.6.0", optional = true }
 fixedbitset = "0.4.0"
