### 2023-12-07 64ac1ea
Bump rkyv from 0.7.40 to 0.7.42
Bumps [rkyv](https://github.com/rkyv/rkyv) from 0.7.40 to 0.7.42.
- [Release notes](https://github.com/rkyv/rkyv/releases)
- [Changelog](https://github.com/rkyv/rkyv/blob/master/release_checklist.md)
- [Commits](https://github.com/rkyv/rkyv/compare/v0.7.40...v0.7.42)

---
updated-dependencies:
- dependency-name: rkyv
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
