### 2023-12-07 939c8ee
Bump conda-incubator/setup-miniconda from 3.0.0 to 3.0.1 (#2860)
Bumps
[conda-incubator/setup-miniconda](https://github.com/conda-incubator/setup-miniconda)
from 3.0.0 to 3.0.1.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/conda-incubator/setup-miniconda/releases">conda-incubator/setup-miniconda's
releases</a>.</em></p>
<blockquote>
<h2>Version 3.0.1</h2>
<h3>Fixes</h3>
<ul>
<li><a
href="https://redirect.github.com/conda-incubator/setup-miniconda/issues/325">#325</a>
Fix environment activation on windows (a v3 regression) due to
hard-coded install 
[…]
