### 2023-12-07 52ef45c
Bump cachix/cachix-action from 12 to 13 (#2861)
Bumps [cachix/cachix-action](https://github.com/cachix/cachix-action)
from 12 to 13.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/cachix/cachix-action/releases">cachix/cachix-action's
releases</a>.</em></p>
<blockquote>
<h2>cachix-action-v13</h2>
<ul>
<li>run on node20</li>
<li>add <code>skipAddingSubstituter</code> to skip running <code>cachix
use</code></li>
</ul>
</blockquote>
</details>
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/cachix/cachix-action/commit/6a2e08b5eb
[…]
