### 2023-12-08 326a747
MRG: Various minor updates to docs (#2854)
Adds FAQ to sidebar. Fixes
https://github.com/sourmash-bio/sourmash/issues/2844.
Adds information about memory usage vs ksize to FAQ. Closes
https://github.com/sourmash-bio/sourmash/issues/2843.
