+++ doc/sidebar.md
@@ -38,6 +38,7 @@ X and Linux. They require about 5 GB of disk space and 5 GB of RAM.
 * [Publications about sourmash](publications.md)
 * [A guide to the internal design and structure of sourmash](sourmash-internals.md)
 * [Funding acknowledgements](funding.md)
+* [Frequently asked questions](faq.md)
 
 ## Developing and extending sourmash
 
