+++ .github/workflows/build_wheel.yml
@@ -39,7 +39,7 @@ jobs:
         with:
           fetch-depth: 0
 
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         name: Install Python
         with:
           python-version: '3.10'
@@ -61,7 +61,7 @@ jobs:
         PYODIDE_VERSION: "0.23.0"
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         with:
            python-version: "3.11.2"
       - run: |
