+++ src/core/Cargo.toml
@@ -45,7 +45,7 @@ murmurhash3 = "0.0.5"
 niffler = { version = "2.3.1", default-features = false, features = [ "gz" ] }
 nohash-hasher = "0.2.0"
 num-iter = "0.1.43"
-once_cell = "1.18.0"
+once_cell = "1.19.0"
 ouroboros = "0.18.0"
 piz = "0.5.0"
 primal-check = "0.3.1"
