### 2023-12-11 0d7c147
Bump ouroboros from 0.18.0 to 0.18.1
Bumps [ouroboros](https://github.com/joshua-maros/ouroboros) from 0.18.0 to 0.18.1.
- [Commits](https://github.com/joshua-maros/ouroboros/commits)

---
updated-dependencies:
- dependency-name: ouroboros
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
