### 2023-12-12 61da777
MRG: bump version to 4.8.6-dev, post-release (#2877)
Please replace this text with:

* a brief description of your changes in this PR
* which issues this fixes in the issue tracker, if any ("Fixes #XXX")
* which issues this PR is related to, if any ("Ref #XXX")

Please also be sure to note here if file formats, command-line
interface, and/or the top-level sourmash API will change because of
this PR.

If you are a new contributor, please add your name and
[your ORCID](https://orcid.org) to the `pyproject.toml` author list
(maintaining alphabetical order by last name).
If you don't have an ORCID
[…]
