### 2023-12-14 a39c9c7
lower macos deployment target on x86_64
