+++ .github/workflows/build_wheel_all_archs.yml
@@ -14,7 +14,7 @@ permissions:
 
 jobs:
 
-  build_wheels:
+  linux:
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -44,11 +44,34 @@ jobs:
           name: wheels
           path: dist
 
+  windows:
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.10'
+          architecture: ${{ matrix.target }}
+
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          target: x64
+          args: --release --out dist --find-interpreter
+          sccache: 'true'
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v3
+        with:
+          name: wheels
+          path: dist
+
   release:
     name: Publish wheels
     runs-on: ubuntu-latest
     if: startsWith(github.ref, 'refs/tags/v')
-    needs: build_wheels
+    needs: [linux, windows]
 
     steps:
       - name: Fetch wheels from artifacts
