### 2023-12-17 94b3631
CI: macos deployment target and maturin updates (#2879)
- macOS fixes from
https://github.com/conda-forge/sourmash-minimal-feedstock/pull/55
- bump maturin to 1.4.0
- use maturin-action for linux `aarch64` and `ppc64le` (`s390x` is still
annoying to get right) and windows (turns out it just works)
