### 2023-12-18 27db282
Bump rkyv from 0.7.42 to 0.7.43 (#2880)
Bumps [rkyv](https://github.com/rkyv/rkyv) from 0.7.42 to 0.7.43.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/rkyv/rkyv/releases">rkyv's
releases</a>.</em></p>
<blockquote>
<h2>0.7.43</h2>
<p>This bugfix release closes a number of issues:</p>
<ul>
<li>Fix documentation typo (<a
href="https://redirect.github.com/rkyv/rkyv/issues/433">#433</a>)</li>
<li>Add a <code>with_capacity</code> constructor for
<code>SharedSerializeMap</code>, <code>SharedDeserializeMap</code>,
<code>DefaultValidator</code>, and <code>S
[…]
