+++ src/core/Cargo.toml
@@ -49,7 +49,7 @@ once_cell = "1.19.0"
 ouroboros = "0.18.1"
 piz = "0.5.0"
 primal-check = "0.3.1"
-rkyv = { version = "0.7.42", optional = true }
+rkyv = { version = "0.7.43", optional = true }
 roaring = "0.10.2"
 rayon = { version = "1.8.0", optional = true }
 serde = { version = "1.0.168", features = ["derive"] }
