+++ .github/workflows/build_wheel.yml
@@ -51,7 +51,7 @@ jobs:
           CIBW_ARCHS_LINUX: ${{ matrix.arch }}
           CIBW_ARCHS_MACOS: ${{ matrix.arch }}
 
-      - uses: actions/upload-artifact@v3
+      - uses: actions/upload-artifact@v4
         with:
           path: './wheelhouse/sourmash*.whl'
 
@@ -79,7 +79,7 @@ jobs:
           export RUSTC_BOOTSTRAP=1
           pyodide build
 
-      - uses: actions/upload-artifact@v3
+      - uses: actions/upload-artifact@v4
         with:
           path: './dist/sourmash*.whl'
 
