+++ .github/workflows/build_wheel_all_archs.yml
@@ -76,7 +76,7 @@ jobs:
     steps:
       - name: Fetch wheels from artifacts
         id: fetch_artifacts
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           path: 'wheels/'
 
