+++ .github/workflows/build_wheel.yml
@@ -93,7 +93,7 @@ jobs:
     steps:
       - name: Fetch wheels from artifacts
         id: fetch_artifacts
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           path: 'wheels/'
 
