### 2023-12-19 172f664
Bump supercharge/redis-github-action from 1.7.0 to 1.8.0 (#2885)
Bumps
[supercharge/redis-github-action](https://github.com/supercharge/redis-github-action)
from 1.7.0 to 1.8.0.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/supercharge/redis-github-action/releases">supercharge/redis-github-action's
releases</a>.</em></p>
<blockquote>
<h2>1.8.0</h2>
<p>Release 1.8.0</p>
</blockquote>
</details>
<details>
<summary>Changelog</summary>
<p><em>Sourced from <a
href="https://github.com/supercharge/redis-github-action/blob/main/CHANGELOG.md">supercharge/red
[…]
