+++ .github/workflows/build_wheel.yml
@@ -53,6 +53,7 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
+          name: binary-${{ matrix.build }}-${{ matrix.os }}
           path: './wheelhouse/sourmash*.whl'
 
   build_wasm:
