### 2023-12-19 58e7edf
standardize artifact prefix, fix names for win and wasm
