+++ .github/workflows/build_wheel_all_archs.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
-          name: wheels-${{ matrix.target }}
+          name: wheel-${{ matrix.target }}
           path: dist
 
   windows:
@@ -64,7 +64,7 @@ jobs:
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
-          name: wheels
+          name: wheel-win64
           path: dist
 
   release:
