### 2023-12-20 895dc15
MRG: fix uploading of wheels after upload-artifact upgrade. (#2887)
This fixes a problem introduced by
https://github.com/sourmash-bio/sourmash/pull/2883, via the solution
mentioned
[here](https://github.com/actions/upload-artifact#not-uploading-to-the-same-artifact)
of naming artifacts using matrix variables.

The output artifacts from the updated `build_wheel.yml` are:
<img width="1062" alt="Screenshot 2023-12-19 at 1 41 24 PM"
src="https://github.com/sourmash-bio/sourmash/assets/51016/55bed1f6-e2e0-40ba-8d66-f6216078256e">

The output artifacts from the updated `build_wheel_all_archs.yml`:

[…]
