### 2023-12-25 703c70e
Bump memmap2 from 0.9.2 to 0.9.3 (#2889)
Bumps [memmap2](https://github.com/RazrFalcon/memmap2-rs) from 0.9.2 to
0.9.3.
<details>
<summary>Changelog</summary>
<p><em>Sourced from <a
href="https://github.com/RazrFalcon/memmap2-rs/blob/master/CHANGELOG.md">memmap2's
changelog</a>.</em></p>
<blockquote>
<h2>[0.9.3] - 2023-12-19</h2>
<h3>Fixed</h3>
<ul>
<li>Build on Android.</li>
</ul>
</blockquote>
</details>
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/RazrFalcon/memmap2-rs/commit/5a49e8ab0bda97770a6ed972633dfb862e17cd38"><code>5a49e8a</code></a>
Version bump.</li>
<
[…]
