### 2023-12-29 sourmash-bio/sourmash#2890 @ctb
address pyopensci review comments for sourmash v4
our reviews are in over at https://github.com/pyOpenSci/software-submission/issues/129!!!

this is a placeholder that will (eventually) contain a checklist for addressing all of the reviewer comments.
