+++ src/core/Cargo.toml
@@ -53,7 +53,7 @@ rkyv = { version = "0.7.43", optional = true }
 roaring = "0.10.2"
 rayon = { version = "1.8.0", optional = true }
 serde = { version = "1.0.168", features = ["derive"] }
-serde_json = "1.0.108"
+serde_json = "1.0.109"
 thiserror = "1.0"
 twox-hash = "1.6.0"
 typed-builder = "0.18.0"
