+++ src/core/Cargo.toml
@@ -64,7 +64,7 @@ criterion = "0.5.1"
 needletail = { version = "0.5.1", default-features = false }
 proptest = { version = "1.4.0", default-features = false, features = ["std"]}
 rand = "0.8.2"
-tempfile = "3.8.1"
+tempfile = "3.9.0"
 
 [[bench]]
 name = "compute"
