### 2024-01-02 f5f7eb3
Bump ouroboros from 0.18.1 to 0.18.2 (#2894)
[//]: # (dependabot-start)
⚠️  **Dependabot is rebasing this PR** ⚠️ 

Rebasing might not happen immediately, so don't worry if this takes some
time.

Note: if you make any changes to this PR yourself, they will take
precedence over the rebase.

---

[//]: # (dependabot-end)

Bumps [ouroboros](https://github.com/joshua-maros/ouroboros) from 0.18.1
to 0.18.2.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/joshua-maros/ouroboros/commits">compare
view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibi
[…]
