### 2024-01-02 b6b264a
update FAQ answer on k-mer size
