+++ src/core/Cargo.toml
@@ -52,7 +52,7 @@ primal-check = "0.3.1"
 rkyv = { version = "0.7.43", optional = true }
 roaring = "0.10.2"
 rayon = { version = "1.8.0", optional = true }
-serde = { version = "1.0.168", features = ["derive"] }
+serde = { version = "1.0.195", features = ["derive"] }
 serde_json = "1.0.110"
 thiserror = "1.0"
 twox-hash = "1.6.0"
