### 2024-01-13 sourmash-bio/sourmash#2905 @ctb
update nix flake & envrc
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982) -

>The nix flake and envrc needs to be updated to make sure the entire environment is available with a simple call to `direnv allow`.
