### 2024-01-13 sourmash-bio/sourmash#2906 @ctb
fix unused imports in benchmarks/benchmarks.py
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)

neither `os` nor `Path` are used, despite being imported.
