### 2024-01-13 sourmash-bio/sourmash#2907 @ctb
examine test running in nix flake
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)

>Not all tests ran out of the box in the dev environment via the nix flake but were easily made to work and seem to work in the GitHub actions.
