### 2024-01-13 sourmash-bio/sourmash#2908 @ctb
consider running flake8 and automatically fixing
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)

>Flake8 revealed a lot of errors, it might not be the worst idea to run code through a linter and fix everything that can be automatically fixed, this may help with the unused imports as well. These are none critical but it helps.
