### 2024-01-13 sourmash-bio/sourmash#2909 @ctb
add repostatus badge to README
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)

we are missing:
>A [repostatus.org](https://www.repostatus.org/) badge,
