### 2024-01-13 sourmash-bio/sourmash#2910 @ctb
add badges for python versions supported
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)
