### 2024-01-13 sourmash-bio/sourmash#2911 @ctb
add comparison / references to other packages to README
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)

> If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.
