### 2024-01-13 sourmash-bio/sourmash#2912 @ctb
for JOSS paper, add references
per [review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1870572982)

 **References:** With DOIs for all those that have one (e.g. papers, datasets, software).
