### 2024-01-13 #129 @ctb
a tracking update - not yet a response ;) - because life does not proceed without github issues!

@elais review issues:

> * The nix flake and envrc needs to be updated to make sure the entire environment is available with a simple call to `direnv allow`.

https://github.com/sourmash-bio/sourmash/issues/2905

> * There are unused imports in benchmarks/benchmarks.py

https://github.com/sourmash-bio/sourmash/issues/2906

> * Not all tests ran out of the box in the dev environment via the nix flake but were easily made to work and seem to work in the GitHub actions.

https://github.com/sourmash-bio/sourmash/issues/2907

> * Flake8 revealed a lot of errors, it might not be the worst idea to run code through a linter and fix everything that can be automatically fixed, this may help with the unused imports as well. These are none critical but it helps.
> * [ ]  Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

https://github.com/sourmash-bio/sourmash/issues/2908

> * [ ]  A [repostatus.org](https://www.repostatus.org/) badge,

https://github.com/sourmash-bio/sourmash/issues/2909

> * [ ]  Python versions supported,

https://github.com/sourmash-bio/sourmash/issues/2910

> If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.

https://github.com/sourmash-bio/sourmash/issues/2911

>     * [ ]  **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

https://github.com/sourmash-bio/sourmash/issues/2912
