### 2024-01-13 sourmash-bio/sourmash#2913 @ctb
update benchmarks.py "magic numbers"
per [@LilyAnderssonLee! review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

>Magic numbers: There are some magic numbers in the code (e.g., 500, 21, 3000, 10000, 1000, 20). Is it possible to define these numbers as constants with meaningful names or add the comments to make the code more self-explanatory?
