### 2024-01-13 sourmash-bio/sourmash#2914 @ctb
fix download link in tutorial
per [@LilyAnderssonLee! review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

>he download link doesn't work in the section [#Building your own LCA database](https://sourmash.readthedocs.io/en/latest/tutorials-lca.html). The correct link should be `curl -L https://osf.io/bw8d7/download -o delmont-subsample-sigs.tar.gz`
