### 2024-01-13 sourmash-bio/sourmash#2915 @ctb
allow users to adjust labels, etc. in plot_dendrogram function
per [@LilyAnderssonLee! review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

> Is it possible to include arguments to allow users to adjust the legends of both matrix and dendrogram plots in the function mentioned in the example: `sourmash plot --pdf --lables ecoli_cmp` ?
