### 2024-01-13 sourmash-bio/sourmash#2916 @ctb
fix usage of `mamba` command in tutorial
per [@LilyAnderssonLee review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

>`mamba` was not callable after installing `miniforge3` on my Mac OS.
