### 2024-01-13 sourmash-bio/sourmash#2917 @ctb
refactor duplicate code in test_sourmash.py
per [@LilyAnderssonLee review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

>Refactoring duplicate code into helper functions to avoid redundancy and improve maintainability.
