### 2024-01-13 sourmash-bio/sourmash#2918 @ctb
provide/update/refactor documentation around k-mer sizes and scaled
per [@LilyAnderssonLee review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

>I may have overlooked it in your documentation. It would be helpful if you could provide some recommendations for the selection of `k-mers` and `scales`.

and

>What are the consequences of using different scales? Do you recommend using the same scale to generate the signatures for genomes or databases?
