### 2024-01-13 sourmash-bio/sourmash#2919 @ctb
add explanations & default value to `sourmash lca classify -h`
per [@LilyAnderssonLee review](https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1838233643)

>Is it possible to add explanations and default values to `--threshold` in the help message of `sourmash lca classify -h`
