### 2024-01-13 #129 @ctb
tracking links for @LilyAnderssonLee review -
>     * [ ]  Badges for:
>       * [ ]  A [repostatus.org](https://www.repostatus.org/) badge,
>       * [ ]  Python versions supported,
> _NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the a badge for pyOpenSci peer-review will be provided upon acceptance.)_

* https://github.com/sourmash-bio/sourmash/issues/2909
* https://github.com/sourmash-bio/sourmash/issues/2910

>     * [ ]  If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.

* https://github.com/sourmash-bio/sourmash/issues/2911

>     * [ ]  **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide).
>       * [ ]  Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

* https://github.com/sourmash-bio/sourmash/issues/2908

> #### For packages also submitting to JOSS
>     * [ ]  **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

* https://github.com/sourmash-bio/sourmash/issues/2912

>     * 1. `mamba` was not callable after installing `miniforge3` on my Mac OS.

* https://github.com/sourmash-bio/sourmash/issues/2916

>     * 2. Is it possible to include arguments to allow users to adjust the legends of both matrix and dendrogram plots in the function mentioned in the example: `sourmash plot --pdf --lables ecoli_cmp` ?

* https://github.com/sourmash-bio/sourmash/issues/2915

>     * 3. Regarding `benchmarks/benchmarks.py` :
>
>          * 1. Unused libraries: `os` & `Path`
>          * 2. Magic numbers: There are some magic numbers in the code (e.g., 500, 21, 3000, 10000, 1000, 20). Is it possible to define these numbers as constants with meaningful names or add the comments to make the code more self-explanatory?

* https://github.com/sourmash-bio/sourmash/issues/2906
* https://github.com/sourmash-bio/sourmash/issues/2913

>     * 4. The download link doesn't work in the section [#Building your own LCA database](https://sourmash.readthedocs.io/en/latest/tutorials-lca.html). The correct link should be `curl -L https://osf.io/bw8d7/download -o delmont-subsample-sigs.tar.gz`

* https://github.com/sourmash-bio/sourmash/issues/2914

>     * 4. Regarding `test/test_sourmash.py` : Refactoring duplicate code into helper functions to avoid redundancy and improve maintainability.

https://github.com/sourmash-bio/sourmash/issues/2917

>     * 5. I may have overlooked it in your documentation. It would be helpful if you could provide some recommendations for the selection of `k-mers` and `scales`.
>     * 6. What are the consequences of using different scales? Do you recommend using the same scale to generate the signatures for genomes or databases?

>
>
> ```
> sourmash gather ecoli-genome.sig inter/genbank-k31.lca.json.gz
> WARNING: final scaled was 10000, vs query scaled of 1000
> ```

* https://github.com/sourmash-bio/sourmash/issues/2918

 >     * 7. Is it possible to add explanations and default values to `--threshold` in the help message of `sourmash lca classify -h`

* https://github.com/sourmash-bio/sourmash/issues/2919
