### 2024-01-14 sourmash-bio/sourmash#2908 @luizirber
time for https://github.com/sourmash-bio/sourmash/pull/2427 to be updated?
