### 2024-01-14 #129 @luizirber
> * The nix flake and envrc needs to be updated to make sure the entire environment is available with a simple call to `direnv allow`.

re https://github.com/sourmash-bio/sourmash/issues/2905: @elais, are you per chance running this on macOS? This works on Linux and I never really tried on macOS because I don't have access to one to try it out, but I can try to fix it too
