### 2024-01-14 sourmash-bio/sourmash#2908 @ctb
I knew that PR was out there somewhere... yes, I think so. You up for it?
