### 2024-01-14 sourmash-bio/sourmash#2908 @luizirber
Yup
