### 2024-01-14 sourmash-bio/sourmash#2916 @ctb
looks like `tutorial-install.md` is the culprit here. need to fix that, probably by switching to mambaforge? Or whatever is recommended these days.
