### 2024-01-14 sourmash-bio/sourmash#2915 @ctb
ref https://github.com/sourmash-bio/sourmash/issues/2587, https://github.com/sourmash-bio/sourmash/issues/2452

I think this might be an easy fix, just gotta dig in to what the more recently-added options are here.
