### 2024-01-14 sourmash-bio/sourmash#2918 @ctb
Partially answered here https://sourmash.readthedocs.io/en/latest/using-sourmash-a-guide.html#what-k-mer-size-s-should-i-use and here https://sourmash.readthedocs.io/en/latest/using-sourmash-a-guide.html#what-resolution-should-my-signatures-be-how-should-i-create-them. I think we should add a pointer to this in the FAQ to make it more discoverable, and also highlight this section:

>One additional wrinkle is that we provide a number of [precalculated databases](https://sourmash.readthedocs.io/en/latest/databases.html) at k=21, k=31, and k=51. It is often convenient to calculate signatures at these sizes so that you can use these databases.
