### 2024-01-14 sourmash-bio/sourmash#2917 @mr-eyes
I have yet to try copilot in writing tests, but overall, it's super challenging for the copilot (at least for now) to understand the testing context and generate the unit tests like the ones on sourmash. However, copilot might be able to modernize some tests. **TLDR** I don't recommend using copilot in rewriting/generating tests specifically for sourmash, it's too complicated.
