### 2024-01-15 0a6444a
update magic numbers in benchmarks
