+++ benchmarks/benchmarks.py
@@ -1,28 +1,46 @@
-import os
 import random
-from pathlib import Path
 from tempfile import NamedTemporaryFile
-
 
 from sourmash.sbt_storage import ZipStorage
 from sourmash.minhash import MinHash
+
+RANDOM_SEQ_SIZE=3000
+RANDOM_SEQ_SAMPLE=300
+
+MINHASH_NUM=500
+MINHASH_K=21
+
+GET_MINS_RANGE=500
+ADD_HASH_RANGE=10_000
+ADD_MANY_RANGE=1000
+SIMILARITY_TIMES=500
+COUNT_COMMON_TIMES=500
+MERGE_TIMES=500
+COPY_TIMES=500
+CONCAT_TIMES=500
+SET_ABUNDANCES_RANGE=500
+ZIP_STORAGE_WRITE=100_000
+ZIP_STORAGE_LOAD=20
 
 
 def load_sequences():
     sequences = []
     for i in range(10):
-        random_seq = random.sample("A,C,G,T".split(",") * 3000, 300)
+        random_seq = random.sample("A,C,G,T".split(",") * RANDOM_SEQ_SIZE,
+                                   RANDOM_SEQ_NUM)
         sequences.append("".join(random_seq))
     return sequences
 
 
 class TimeMinHashSuite:
     def setup(self):
-        self.mh = MinHash(500, 21, track_abundance=False)
-        self.protein_mh = MinHash(500, 21, is_protein=True, track_abundance=False)
+        self.mh = MinHash(MINHASH_NUM, MINHASH_K, track_abundance=False)
+        self.protein_mh = MinHash(MINHASH_NUM, MINHASH_K, is_protein=True,
+                                  track_abundance=False)
         self.sequences = load_sequences()
 
-        self.populated_mh = MinHash(500, 21, track_abundance=False)
+        self.populated_mh = MinHash(MINHASH_NUM, MINHASH_K,
+                                    track_abundance=False)
         for seq in self.sequences:
             self.populated_mh.add_sequence(seq)
 
@@ -40,52 +58,53 @@ class TimeMinHashSuite:
 
     def time_get_mins(self):
         mh = self.populated_mh
-        for i in range(500):
+        for i in range(GET_MINS_RANGE):
             mh.get_mins()
 
     def time_add_hash(self):
         mh = self.mh
-        for i in range(10000):
+        for i in range(ADD_HASH_RANGE):
             mh.add_hash(i)
 
     def time_add_many(self):
         mh = self.mh
-        mh.add_many(list(range(1000)))
+        mh.add_many(list(range(ADD_MANY_RANGE)))
 
     def time_similarity(self):
         mh = self.mh
         other_mh = self.populated_mh
-        for i in range(500):
+        for i in range(SIMILARITY_TIMES):
             mh.similarity(other_mh)
 
     def time_count_common(self):
         mh = self.mh
         other_mh = self.populated_mh
-        for i in range(500):
+        for i in range(COUNT_COMMON_TIMES):
             mh.count_common(other_mh)
 
     def time_merge(self):
         mh = self.mh
         other_mh = self.populated_mh
-        for i in range(500):
+        for i in range(MERGE_TIMES):
             mh.merge(other_mh)
 
     def time_copy(self):
         mh = self.populated_mh
-        for i in range(500):
+        for i in range(COPY_TIMES):
             mh.__copy__()
 
     def time_concat(self):
         mh = self.mh
         other_mh = self.populated_mh
-        for i in range(500):
+        for i in range(CONCAT_TIMES):
             mh += other_mh
 
 
 class PeakmemMinHashSuite:
     def setup(self):
-        self.mh = MinHash(500, 21, track_abundance=True)
-        self.protein_mh = MinHash(500, 21, is_protein=True, track_abundance=True)
[…]
