### 2024-01-15 5e3e3ec
add scaled FAQ, adjust ksize answer
