### 2024-01-15 4f32abc
MRG: fix benchmark code & a few other small issues from pyOpenSci review (#2920)
Address issues:
* https://github.com/sourmash-bio/sourmash/issues/2919
* https://github.com/sourmash-bio/sourmash/issues/2913
* https://github.com/sourmash-bio/sourmash/issues/2906
* https://github.com/sourmash-bio/sourmash/issues/2914

Fixes https://github.com/sourmash-bio/sourmash/issues/2919
Fixes https://github.com/sourmash-bio/sourmash/issues/2913
Fixes https://github.com/sourmash-bio/sourmash/issues/2906
Fixes https://github.com/sourmash-bio/sourmash/issues/2914

Remaining TODO:
* test `benchmarks/benchmarks
[…]
