### 2024-01-15 df57a38
Update doc/faq.md
Co-authored-by: Colton Baumler <63077899+ccbaumler@users.noreply.github.com>
