### 2024-01-15 11af4d5
MRG: add scaled FAQ, adjust ksize answer (#2921)
Adds the following FAQ entry to address
https://github.com/sourmash-bio/sourmash/issues/2918:

> ## What scaled values should I use with sourmash?
> 
> We recommend scaled=1000 or scaled=10000 when working with bacterial
> and archaeal sketches and DNA. We have quite a bit of experience with
> this, and even some
> [published
benchmarks](https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-022-05103-0)
> showing that this works very well.  You may need to use lower scaled
> values with smaller query and target sequences, such as v
[…]
