+++ pyproject.toml
@@ -114,7 +114,7 @@ demo = [
 doc = [
   "sphinx>=4.4.0,<8",
   "myst-parser==2.0.0",
-  "Jinja2==3.1.2",
+  "Jinja2==3.1.3",
   "alabaster",
   "sphinxcontrib-napoleon",
   "nbsphinx",
