### 2024-01-15 dac86fd
Bump histogram from 0.8.3 to 0.8.4
Bumps [histogram](https://github.com/pelikan-io/rustcommon) from 0.8.3 to 0.8.4.
- [Commits](https://github.com/pelikan-io/rustcommon/commits)

---
updated-dependencies:
- dependency-name: histogram
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
