### 2024-01-16 1dd6711
Bump wasm-bindgen from 0.2.89 to 0.2.90 (#2925)
Bumps [wasm-bindgen](https://github.com/rustwasm/wasm-bindgen) from
0.2.89 to 0.2.90.
<details>
<summary>Changelog</summary>
<p><em>Sourced from <a
href="https://github.com/rustwasm/wasm-bindgen/blob/main/CHANGELOG.md">wasm-bindgen's
changelog</a>.</em></p>
<blockquote>
<h2><a
href="https://github.com/rustwasm/wasm-bindgen/compare/0.2.89...0.2.90">0.2.90</a></h2>
<p>Released 2024-01-06</p>
<h3>Fixed</h3>
<ul>
<li>Fix JS shim default path detection for the no-modules target.
<a
href="https://redirect.github.com/rustwasm/wasm-bindgen/pull/3748">#37
[…]
