### 2024-01-16 ee55364
Bump cachix/cachix-action from 13 to 14 (#2926)
Bumps [cachix/cachix-action](https://github.com/cachix/cachix-action)
from 13 to 14.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/cachix/cachix-action/releases">cachix/cachix-action's
releases</a>.</em></p>
<blockquote>
<h2>cachix-action-v14</h2>
<p><a
href="https://blog.cachix.org/posts/2024-01-12-cachix-v1-7/">https://blog.cachix.org/posts/2024-01-12-cachix-v1-7/</a></p>
</blockquote>
</details>
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/cachix/cachix-action/commit/18c
[…]
