+++ .github/workflows/dev_envs.yml
@@ -16,7 +16,7 @@ jobs:
         extra_nix_config: |
           access-tokens = github.com=${{ secrets.GITHUB_TOKEN }}
 
-    - uses: cachix/cachix-action@v13
+    - uses: cachix/cachix-action@v14
       with:
         name: sourmash-bio
         authToken: '${{ secrets.CACHIX_AUTH_TOKEN }}'
