### 2024-01-16 bc00e28
Bump cachix/install-nix-action from 24 to 25 (#2927)
Bumps
[cachix/install-nix-action](https://github.com/cachix/install-nix-action)
from 24 to 25.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/cachix/install-nix-action/releases">cachix/install-nix-action's
releases</a>.</em></p>
<blockquote>
<h2>install-nix-action-v25</h2>
<p>Nix 2.19.2</p>
</blockquote>
</details>
<details>
<summary>Commits</summary>
<ul>
<li><a
href="https://github.com/cachix/install-nix-action/commit/6004951b182f8860210c8d6f0d808ec5b1a33d28"><code>6004951</code></a>
Nix: 2.19.1 
[…]
