### 2024-01-16 c54c2e8
adjust order of badges
