### 2024-01-16 sourmash-bio/sourmash#2912 @bluegenes
Note paper references are currently in a separate file from the paper: https://github.com/sourmash-bio/sourmash/blob/latest/paper.bib
