### 2024-01-17 641d119
downsample within select
