+++ src/core/src/storage.rs
@@ -432,10 +432,9 @@ impl SigStore {
     }
 }
 
-
 impl Select for SigStore {
     fn select(mut self, selection: &Selection) -> Result<Self> {
-        // TODO: find better error (perhaps DataNotFound or similar?)
+        // TODO: find better error (perhaps ReadDataError or similar?)
         let sig = self.data.take().ok_or(Error::MismatchKSizes)?;
 
         // first, select based on ksize, compatible scaled
@@ -445,7 +444,7 @@ impl Select for SigStore {
             for sketch in selected.iter_mut() {
                 if let Sketch::MinHash(mh) = sketch {
                     let sig_scaled = mh.scaled() as u32;
-                    if sig_scaled != sel_scaled && sig_scaled < sel_scaled {
+                    if sig_scaled < sel_scaled {
                         // downsample in place
                         mh.downsample_scaled(sel_scaled as u64)?;
                     }
