### 2024-01-17 869c7c4
move ds logic back to Signature Select
