### 2024-01-18 1e626f2
actually replace sketch with ds version; add test
