### 2024-01-18 986fc3d
consolidate tests with other signature tests
