### 2024-01-20 2e1775d
switch to suggesting conda-forge and sourmash-minimal in README
