### 2024-01-22 #129 @elais
Yes I am running it on an M1 Mac. I could try running it in a virtual machine to see if I get better results.
