### 2024-01-22 ad3dfdf
remove research notice
