+++ README.md
@@ -116,19 +116,5 @@ Tests require py.test and can be run with `make test`.
 Please see [the developer notes](doc/developer.md) for more information
 on getting set up with a development environment.
 
-## Research notice
-
-Please note that this repository is participating in a study into sustainability
- of open source projects. Data will be gathered about this repository for
- approximately the next 12 months, starting from 2021-06-11.
-
-Data collected will include number of contributors, number of PRs, time taken to
- close/merge these PRs, and issues closed.
-
-For more information, please visit
-[our informational page](https://sustainable-open-science-and-software.github.io/) or download our [participant information sheet](https://sustainable-open-science-and-software.github.io/assets/PIS_sustainable_software.pdf).
-
-----
-
 CTB
 Jan 2024
