### 2024-01-22 b5c4f1b
MRG: update README with repostatus and pyver badges, and Windows support (#2928)
Direct link to updated README:
[link](https://github.com/sourmash-bio/sourmash/blob/update_readme/README.md)

Adds repostatus badge as well as list of python versions supported.

Switches to recommending conda-forge and sourmash-minimal, since that
supports Windows ref #1923.

Also adds direct mention of operating systems supported, including
Windows (ref https://github.com/sourmash-bio/sourmash/issues/1923).

Fixes https://github.com/sourmash-bio/sourmash/issues/2909
Fixes https://github.com/sourmash-bio/sourmash
[…]
