+++ .github/workflows/python.yml
@@ -33,7 +33,7 @@ jobs:
           echo "::set-output name=dir::$(pip cache dir)"
 
       - name: pip cache
-        uses: actions/cache@v3
+        uses: actions/cache@v4
         with:
           path: ${{ steps.pip-cache.outputs.dir }}
           key: ${{ runner.os }}-pip-v4-${{ hashFiles('**/pyproject.toml') }}
@@ -62,7 +62,7 @@ jobs:
           redis-version: 6
 
       - name: tox cache
-        uses: actions/cache@v3
+        uses: actions/cache@v4
         with:
           path: .tox/
           key: ${{ runner.os }}-tox-v4-${{ hashFiles('**/pyproject.toml') }}
