### 2024-01-23 454a614
roll back byteorder for sourmash_plugin_branchwater (zip) compatibility
