### 2024-01-23 efd1ee4
also roll back once_cell and wasm_bindgen
