### 2024-01-23 93ec3a8
Bump histogram from 0.8.4 to 0.9.0 (#2935)
Bumps [histogram](https://github.com/pelikan-io/rustcommon) from 0.8.4
to 0.9.0.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/pelikan-io/rustcommon/commits">compare
view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibility
score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=histogram&package-manager=cargo&previous-version=0.8.4&new-version=0.9.0)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-updates#about-compa
[…]
