+++ .github/workflows/build_wheel.yml
@@ -69,7 +69,7 @@ jobs:
           pip install pyodide-build==${PYODIDE_VERSION} "pydantic<2"
           pyodide config get emscripten_version  # trigger setup
           echo EMSCRIPTEN_VERSION=$(pyodide config get emscripten_version) >> $GITHUB_ENV
-      - uses: mymindstorm/setup-emsdk@v13
+      - uses: mymindstorm/setup-emsdk@v14
         with:
            version: ${{ env.EMSCRIPTEN_VERSION }}
       - uses: actions-rs/toolchain@v1
