### 2024-01-23 dee4dc9
Bump ouroboros from 0.18.2 to 0.18.3 (#2936)
Bumps [ouroboros](https://github.com/someguynamedjosh/ouroboros) from
0.18.2 to 0.18.3.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/someguynamedjosh/ouroboros/commits">compare
view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibility
score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=ouroboros&package-manager=cargo&previous-version=0.18.2&new-version=0.18.3)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-upd
[…]
