### 2024-01-23 f8f7f5c
Bump rayon from 1.8.0 to 1.8.1 (#2937)
Bumps [rayon](https://github.com/rayon-rs/rayon) from 1.8.0 to 1.8.1.
<details>
<summary>Changelog</summary>
<p><em>Sourced from <a
href="https://github.com/rayon-rs/rayon/blob/master/RELEASES.md">rayon's
changelog</a>.</em></p>
<blockquote>
<h1>Release rayon 1.8.1 / rayon-core 1.12.1 (2024-01-17)</h1>
<ul>
<li>The new <code>&quot;web_spin_lock&quot;</code> crate feature makes
mutexes spin on the main
browser thread in WebAssembly, rather than suffer an error about
forbidden
<code>atomics.wait</code> if they were to block in that context. Thanks
<a
href="
[…]
