+++ src/core/Cargo.toml
@@ -51,7 +51,7 @@ piz = "0.5.0"
 primal-check = "0.3.1"
 rkyv = { version = "0.7.43", optional = true }
 roaring = "0.10.2"
-rayon = { version = "1.8.0", optional = true }
+rayon = { version = "1.8.1", optional = true }
 serde = { version = "1.0.195", features = ["derive"] }
 serde_json = "1.0.111"
 thiserror = "1.0"
