### 2024-01-23 4f6b349
Bump wasm-bindgen-test from 0.3.39 to 0.3.40
Bumps [wasm-bindgen-test](https://github.com/rustwasm/wasm-bindgen) from 0.3.39 to 0.3.40.
- [Release notes](https://github.com/rustwasm/wasm-bindgen/releases)
- [Changelog](https://github.com/rustwasm/wasm-bindgen/blob/main/CHANGELOG.md)
- [Commits](https://github.com/rustwasm/wasm-bindgen/commits)

---
updated-dependencies:
- dependency-name: wasm-bindgen-test
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
