### 2024-01-23 a0f3016
Bump web-sys from 0.3.66 to 0.3.67 (#2939)
Bumps [web-sys](https://github.com/rustwasm/wasm-bindgen) from 0.3.66 to
0.3.67.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/rustwasm/wasm-bindgen/commits">compare
view</a></li>
</ul>
</details>
<br />


[![Dependabot compatibility
score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=web-sys&package-manager=cargo&previous-version=0.3.66&new-version=0.3.67)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-updates#about-compa
[…]
