+++ src/core/Cargo.toml
@@ -92,7 +92,7 @@ version = "0.2.90"
 features = ["serde-serialize"]
 
 [target.'cfg(all(target_arch = "wasm32", target_os="unknown"))'.dependencies.web-sys]
-version = "0.3.66"
+version = "0.3.67"
 features = ["console", "File"]
 
 [target.'cfg(all(target_arch = "wasm32"))'.dependencies.chrono]
