### 2024-01-23 94b88cc
MRG: in core, enable downsample within select (#2931)
Attempting #1292 in order to move forward
https://github.com/sourmash-bio/sourmash_plugin_branchwater/pull/134

Modifies `Signature` `Select` to downsample automatically.

- for scaled sketches, while checking ksize, we also retain only
sketches that have the right scaled or can be downsampled (scaled <=
selection.scaled())
- next, we iterate through the sketches and downsample any where scaled
< selection.scaled()

Note that for `sourmash_plugin_branchwater` compatibility, we need:

- `byteorder` = "1.4.3"
- `wasm-bindgen` = "0.2.89"
- `on
[…]
