+++ src/core/src/signature.rs
@@ -771,8 +771,14 @@ impl Select for Signature {
             } else {
                 valid
             };
-            // TODO: execute downsample if needed
-
+            // keep compatible scaled if applicable
+            if let Some(sel_scaled) = selection.scaled() {
+                valid = if let Sketch::MinHash(mh) = s {
+                    valid && mh.scaled() <= sel_scaled as u64
+                } else {
+                    valid
+                };
+            }
             /*
             valid = if let Some(abund) = selection.abund() {
                 valid && *s.with_abundance() == abund
@@ -785,8 +791,20 @@ impl Select for Signature {
                 valid
             };
             */
+
             valid
         });
+
+        // downsample the retained sketches if needed.
+        if let Some(sel_scaled) = selection.scaled() {
+            for sketch in self.signatures.iter_mut() {
+                if let Sketch::MinHash(mh) = sketch {
+                    if (mh.scaled() as u32) < sel_scaled {
+                        *sketch = Sketch::MinHash(mh.downsample_scaled(sel_scaled as u64)?);
+                    }
+                }
+            }
+        }
         Ok(self)
     }
 }
@@ -840,6 +858,10 @@ mod test {
     use crate::signature::SigsTrait;
 
     use super::Signature;
+
+    use crate::prelude::Select;
+    use crate::selection::Selection;
+    use crate::sketch::Sketch;
 
     #[test]
     fn load_sig() {
@@ -977,6 +999,49 @@ mod test {
         assert_eq!(sketches.len(), 12);
         for sk in sketches {
             assert_eq!(sk.size(), 500);
+        }
+    }
+
+    #[test]
+    fn selection_with_downsample() {
+        let mut filename = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+        filename.push("../../tests/test-data/47+63-multisig.sig");
+
+        let file = File::open(filename).unwrap();
+        let reader = BufReader::new(file);
+        let sigs: Vec<Signature> = serde_json::from_reader(reader).expect("Loading error");
+
+        // create Selection object
+        let mut selection = Selection::default();
+        selection.set_scaled(2000);
+        // iterate and check scaled
+        for sig in &sigs {
+            let modified_sig = sig.clone().select(&selection).unwrap();
+            for sketch in modified_sig.sketches() {
+                if let Sketch::MinHash(mh) = sketch {
+                    dbg!("scaled: {:?}", mh.scaled());
+                    assert_eq!(mh.scaled(), 2000);
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn selection_scaled_too_low() {
+        let mut filename = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+        filename.push("../../tests/test-data/47+63-multisig.sig");
+
+        let file = File::open(filename).unwrap();
+        let reader = BufReader::new(file);
+        let sigs: Vec<Signature> = serde_json::from_reader(reader).expect("Loading error");
+
+        // create Selection object
+        let mut selection = Selection::default();
+        selection.set_scaled(100);
+        // iterate and check no sigs are returned (original scaled is 1000)
+        for sig in &sigs {
+            let modified_sig = sig.clone().select(&selection).unwrap();
+            assert_eq!(modified_sig.size(), 0);
         }
     }
 }
