### 2024-01-24 bdec5e2
init gather col additions
