### 2024-01-25 c7038be
load sig directly from record, which we already have w/ iter, iter_mut
