### 2024-01-27 23b603b
add potential new fns
