+++ src/core/src/selection.rs
@@ -4,6 +4,7 @@ use typed_builder::TypedBuilder;
 use crate::encodings::HashFunctions;
 use crate::manifest::Record;
 use crate::Result;
+use crate::sketch::Sketch;
 
 #[derive(Default, Debug, TypedBuilder, Clone)]
 pub struct Selection {
@@ -27,6 +28,9 @@ pub struct Selection {
 
     #[builder(default, setter(strip_option))]
     picklist: Option<Picklist>,
+
+    #[builder(default, setter(strip_option))]
+    sketchtype: Option<Sketch>,
 }
 
 #[derive(Default, TypedBuilder, CopyGetters, Getters, Setters, Clone, Debug)]
@@ -119,6 +123,10 @@ impl Selection {
         self.picklist = Some(value);
     }
 
+    pub fn set_sketchtype(&mut self, value: Sketch) {
+        self.sketchtype = Some(value);
+    }
+
     pub fn from_record(row: &Record) -> Result<Self> {
         Ok(Self {
             ksize: Some(*row.ksize()),
@@ -128,6 +136,7 @@ impl Selection {
             scaled: None,
             containment: None,
             picklist: None,
+            sketchtype: None,
         })
     }
 }
