### 2024-01-29 1f08406
minor refactoring of gather
