### 2024-01-29 794dc2c
remove unnecessary iteration
