### 2024-01-29 59f4a2b
remove unnecessary set conversion and subtraction
