### 2024-01-29 2edf313
minor cleanups, remove unnec assign
