### 2024-01-29 f0a45d7
add full column descriptions per #2812
