+++ doc/classifying-signatures.md
@@ -533,3 +533,68 @@ figure out across all of the different use cases for gather.  Perhaps
 in the future we'll find a better way to represent all of these
 numbers in a more clear, concise, and interpretable way; in the
 meantime, we welcome your questions and comments!
+
+## Appendix D: column descriptions for `gather` and `prefetch` CSV output.
+
+## `prefetch` CSV output columns
+
+* `intersect_bp` - integer: size of overlap between match and original query, estimated by multiplying the number of overlapping hashes by `scaled`.
+* `jaccard` - float: Jaccard similarity of the two sketches.
+* `max_containment` - float: max of `f_query_match` and `f_match_query`.
+* `f_query_match` - float: the fraction of the query contained by the match
+* `f_match_query` - float: the fraction of the match contained by the query
+* `match_filename` - string: filename the match sketch was loaded from.
+* `match_name` - string: full name of match sketch.
+* `match_md5` - string: truncated md5sum of match sketch (8 char).
+* `match_bp` - integer: size of match, estimated by multiplying the sketch size by scaled.
+* `query_filename` - string: filename the query sketch was loaded from.
+* `query_name` - string: full name of query sketch.
+* `query_md5` - string: truncated md5sum of query sketch (8 char).
+* `query_bp` - integer: size of query, estimated by multiplying the sketch size by scaled.
+* `ksize` - integer: k-mer size for the sketches used in the comparison.
+* `moltype` - string: molecule type of the sketches.
+* `scaled` - integer: scaled value at which the comparison was done.
+* `query_n_hashes` - integer: number of hashes in the query.
+* `query_abundance` - integer: median hash abundance in the sketch, if available (CTB check: if available; median - or is it just true/false?
+* `query_containment_ani` - float: ANI estimated from the query containment in the match.
+* `match_containment_ani` - float: ANI estimated from the match containment in the query.
+* `average_containment_ani` - float: ANI estimated from the average of the query and match containment.
+* `max_containment_ani` - float: ANI estimated from the max containment between query/match.
+`potential_false_negative` - boolean: True if the sketch size(s) were too small to give a reliable ANI estimate. False if ANI estimate is reliable.
+
+## `gather` CSV output columns
+
+Here the _query_ is typically a metagenome, and the _matches_ are one or more genomes that collectively cover the query.
+
+* `unique_intersect_bp` - integer: size of overlap between match and _remaining_ query, estimated by multiplying the number of overlapping hashes by scaled. Rank/order dependent. Does not double count hashes.
+* `intersect_bp` - integer: size of overlap between match and query, estimated by multiplying the number of overlapping hashes by scaled. Independent of rank order and will often double-count hashes.
+* `f_orig_query` - float: the fraction of the original query represented by this match. Approximates the fraction of metagenomic reads that will map to this genome.
+* `f_match` - float: the containment of the match in the query.
+* `f_unique_to_query` - float: the fraction of matching hashes (unweighted) that are unique to this query; rank dependent. Will sum to the fraction of total k-mers (unweighted) that were identified.
+* `f_unique_weighted` - float: the fraction of matching hashes (weighted by multiplicity) that are unique to this query. This will sum to the fraction of total _weighted_ k-mers that were identified. Approximates the fraction of metagenomic reads that will map to this genome _after_ all previous matches at lower (earlier) ranks are mapped.
+* `average_abund` - float: mean abundance of the weighted hashes unique to the intersection. Empty if query does not have abundance. Rank dependent, does not double count.
+* `median_abund` - integer: median abundance of the weighted hashes unique to the intersection. Empty if query has no abundance. Rank dependent, does not double count.
[…]
