### 2024-01-29 10c7b4c
add manifest scaled select, tests:collection loading+downsampling
