+++ src/core/Cargo.toml
@@ -96,7 +96,7 @@ version = "0.3.67"
 features = ["console", "File"]
 
 [target.'cfg(all(target_arch = "wasm32"))'.dependencies.chrono]
-version = "0.4.28"
+version = "0.4.33"
 features = ["wasmbind"]
 
 [target.'cfg(all(target_arch = "wasm32", target_os="unknown"))'.dev-dependencies]
