### 2024-01-30 bbf02a0
test get_sketch
