+++ src/core/src/selection.rs
@@ -3,8 +3,8 @@ use typed_builder::TypedBuilder;
 
 use crate::encodings::HashFunctions;
 use crate::manifest::Record;
-use crate::Result;
 use crate::sketch::Sketch;
+use crate::Result;
 
 #[derive(Default, Debug, TypedBuilder, Clone)]
 pub struct Selection {
@@ -28,9 +28,8 @@ pub struct Selection {
 
     #[builder(default, setter(strip_option))]
     picklist: Option<Picklist>,
-
-    #[builder(default, setter(strip_option))]
-    sketchtype: Option<Sketch>,
+    // #[builder(default, setter(strip_option))]
+    // sketchtype: Option<Sketch>,
 }
 
 #[derive(Default, TypedBuilder, CopyGetters, Getters, Setters, Clone, Debug)]
@@ -123,9 +122,13 @@ impl Selection {
         self.picklist = Some(value);
     }
 
-    pub fn set_sketchtype(&mut self, value: Sketch) {
-        self.sketchtype = Some(value);
-    }
+    // pub fn sketchtype(&self) -> Option<Sketch> {
+    //     self.sketchtype()
+    // }
+
+    // pub fn set_sketchtype(&mut self, value: Sketch) {
+    //     self.sketchtype = Some(value);
+    // }
 
     pub fn from_record(row: &Record) -> Result<Self> {
         Ok(Self {
@@ -136,7 +139,7 @@ impl Selection {
             scaled: None,
             containment: None,
             picklist: None,
-            sketchtype: None,
+            // sketchtype: None,
         })
     }
 }
