### 2024-01-30 7a85f26
manifest from pathlist
