### 2024-01-30 535e51a
Bump pypa/cibuildwheel from 2.16.2 to 2.16.4 (#2960)
Bumps [pypa/cibuildwheel](https://github.com/pypa/cibuildwheel) from
2.16.2 to 2.16.4.
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/pypa/cibuildwheel/releases">pypa/cibuildwheel's
releases</a>.</em></p>
<blockquote>
<h2>v2.16.4</h2>
<p>🛠 Update manylinux pins to upgrade from a problematic PyPy version.
(<a
href="https://redirect.github.com/pypa/cibuildwheel/issues/1737">#1737</a>)</p>
<h2>v2.16.3</h2>
<ul>
<li>🐛 Fix a bug when building from sdist, where relative paths to files
in the package didn
[…]
