### 2024-01-30 9033d6d
MRG: Fix `tax metagenome` to work on gather output created with `--estimate-ani-ci` (#2952)
The `tax metagenome` code errors out with a confusing error message when
gather results containing confidence intervals for the ANI calculations
are output. This PR adds a test and fixes the problem.

## Details

The error message is:
```
ERROR: '/var/folders/6s/_f373w1d6hdfjc2kjstq97s80000gp/T/sourmashtest_rs5l3b23/gather.csv' is missing columns needed for taxonomic summarization. Please run gather with sourmash >= 4.4.
```

and it is caused by `GatherRow` running across the various extra columns
added
[…]
