### 2024-01-30 196b997
change save_matches to work in streaming manner
