### 2024-01-30 3e5d59b
test minhash, get_sketch failures
