+++ src/core/src/signature.rs
@@ -543,14 +543,12 @@ impl Signature {
             }
             return None;
         }
-        for sk in &self.signatures {
-            match sk {
-                Sketch::MinHash(_) | Sketch::LargeMinHash(_) | Sketch::HyperLogLog(_) => {
-                    return Some(sk);
-                }
-            }
-        }
-        None
+        self.signatures.iter().find(|sk| {
+            matches!(
+                sk,
+                Sketch::MinHash(_) | Sketch::LargeMinHash(_) | Sketch::HyperLogLog(_)
+            )
+        })
     }
 
     // return minhash directly
@@ -561,13 +559,13 @@ impl Signature {
             }
             return None;
         }
-        for sk in &self.signatures {
-            // TODO:  also account for Sketch::LargeMinHash?
+        self.signatures.iter().find_map(|sk| {
             if let Sketch::MinHash(mh) = sk {
-                return Some(mh);
+                Some(mh)
+            } else {
+                None
             }
-        }
-        None
+        })
     }
 
     pub fn from_path<P: AsRef<Path>>(path: P) -> Result<Vec<Signature>, Error> {
@@ -808,18 +806,16 @@ impl Select for Signature {
                 valid
             };
             // keep compatible scaled if applicable
-            if let Some(sel_scaled) = selection.scaled() {
-                valid = match s {
+            valid = if let Some(sel_scaled) = selection.scaled() {
+                match s {
                     Sketch::MinHash(mh) => valid && mh.scaled() <= sel_scaled as u64,
-                    //TODO: Test LMH
-                    // Sketch::LargeMinHash(lmh) => {
-                    //     valid && lmh.scaled() <= sel_scaled as u64
-                    // }
-                    _ => {
-                        valid // Handle other sketch types or invalid cases
-                    }
-                };
-            }
+                    // TODO: test LargeMinHash
+                    // Sketch::LargeMinHash(lmh) => valid && lmh.scaled() <= sel_scaled as u64,
+                    _ => valid, // other sketch types or invalid cases
+                }
+            } else {
+                valid // if selection.scaled() is None, keep prior valid
+            };
             /*
             valid = if let Some(abund) = selection.abund() {
                 valid && *s.with_abundance() == abund
@@ -839,21 +835,12 @@ impl Select for Signature {
         // downsample the retained sketches if needed.
         if let Some(sel_scaled) = selection.scaled() {
             for sketch in self.signatures.iter_mut() {
-                match sketch {
-                    Sketch::MinHash(mh) => {
-                        if (mh.scaled() as u32) < sel_scaled {
-                            let new_mh = mh.downsample_scaled(sel_scaled as u64)?;
-                            *sketch = Sketch::MinHash(new_mh);
-                        }
+                // TODO: also account for LargeMinHash
+                if let Sketch::MinHash(mh) = sketch {
+                    if (mh.scaled() as u32) < sel_scaled {
+                        let new_mh = mh.downsample_scaled(sel_scaled as u64)?;
+                        *sketch = Sketch::MinHash(new_mh);
                     }
-                    // TODO: test LMH
-                    // Sketch::LargeMinHash(lmh) => {
-                    //     if (lmh.scaled() as u32) < sel_scaled {
-                    //         let new_lmh = lmh.downsample_scaled(sel_scaled as u64)?;
-                    //         *sketch = Sketch::LargeMinHash(new_lmh);
-                    //     }
-                    // }
-                    _ => {}
                 }
             }
         }
@@ -1091,6 +1078,37 @@ mod test {
             // error
[…]
