### 2024-01-30 06c2fbd
add pathbuf failure tests
