### 2024-01-30 cd765ab
stop saving gather results
