### 2024-01-30 b26e6bf
add test for trying to build manifest from bad sig path
