### 2024-01-30 409aeb4
handle scaled selection for num sigs
