+++ src/core/src/manifest.rs
@@ -202,7 +202,14 @@ impl Select for Manifest {
                 valid
             };
             valid = if let Some(scaled) = selection.scaled() {
-                valid && row.scaled <= scaled as u64
+                // num sigs have row.scaled = 0
+                valid && row.scaled != 0 && row.scaled <= scaled as u64
+            } else {
+                valid
+            };
+            valid = if let Some(num) = selection.num() {
+                // scaled sigs have row.num = 0
+                valid && row.num != 0 && row.num == num
             } else {
                 valid
             };
