### 2024-01-31 459f9e4
add doi where needed; update publications as needed
