### 2024-02-01 sourmash-bio/sourmash#2911 @bluegenes
I think this checkbox specifically pertains to the README, though this could be useful in either/both places.
