### 2024-02-01 sourmash-bio/sourmash#2911 @ctb
oh! oops.

ok, updated title. If you feel inspired, please post some text here and/or set up a PR - I'll get to it sooner or later if you don't 😆
