### 2024-02-01 e2cde66
add info about sourmash vs other kmer programs
