+++ README.md
@@ -36,7 +36,7 @@ to upgrade!
 
 ----
 
-sourmash is a k-mer analysis multitool, and we aim to provide stable, robust programmatic and command-line APIs for a variety of sequence comparisons. Our unique functionality includes:
+sourmash is a k-mer analysis multitool, and we aim to provide stable, robust programmatic and command-line APIs for a variety of sequence comparisons. Some of our special sauce includes:
 - `FracMinHash` sketching, which enables accurate comparisons (including ANI) between data sets of different sizes
 - `sourmash gather`, a combinatorial k-mer approach for more accurate metagenomic profiling
 
@@ -47,7 +47,7 @@ combined with @ctb's love of whiskey.
 ([Sour mash](https://en.wikipedia.org/wiki/Sour_mash) is used in
 making whiskey.)
 
-Primary authors: [C. Titus Brown](mailto:titus@idyll.org) ([@ctb](http://github.com/ctb)) and [Luiz C. Irber, Jr](mailto:sourmash@luizirber.org) ([@luizirber](http://github.com/luizirber)).
+Primary authors: [C. Titus Brown](mailto:titus@idyll.org) ([@ctb](http://github.com/ctb)), [Luiz C. Irber, Jr](mailto:sourmash@luizirber.org) ([@luizirber](http://github.com/luizirber)), and [N. Tessa Pierce-Ward](mailto:tessa@sourmash.bio) ([@bluegenes](http://github.com/bluegenes)).
 
 sourmash was initially developed by the
 [Lab for Data-Intensive Biology](http://ivory.idyll.org/lab/) at the
