### 2024-02-02 c0ca449
MRG: update readme with maintainers & sourmash comparison info (#2965)
Closes #2911.

Since the world of k-mer software is pretty fast-evolving, I thought it
might be better to just discuss some of our special functionality here
rather than try to compare against a litany of programs.

This PR also updates README.md and `pyproject.toml` with @bluegenes as
maintainer!

---------

Co-authored-by: Luiz Irber <luizirber@users.noreply.github.com>
Co-authored-by: C. Titus Brown <titus@idyll.org>
