### 2024-02-03 1692eb7
MRG: update JOSS paper per pyopensci review (#2964)
fixes https://github.com/sourmash-bio/sourmash/issues/2912

- adds DOIs to references where they were missing
- updates publication information for Rahman Hera (ANI) paper
- adds taxonomic and functional profiling references
- set line length to 80 per markdown specifications

Note, citations are >450 if we include all papers (364 with just
original JOSS + F100).

---------

Co-authored-by: C. Titus Brown <titus@idyll.org>
