### 2024-02-05 97596f5
use Option::iter to simplify code in merge
