### 2024-02-05 sourmash-bio/sourmash#2916 @ctb
Actually, the problem was probably MacOS and the use of zsh instead of bash? The tutorial installs things properly on Linux, and properly directs people to use the install instructions for miniforge on other OSes.
