### 2024-02-05 sourmash-bio/sourmash#2916 @ctb
(but still, this inspired https://github.com/sourmash-bio/sourmash/pull/2972 :)
