### 2024-02-05 sourmash-bio/sourmash#2915 @ctb
looks like https://github.com/sourmash-bio/sourmash/pull/2598 is the fix! Now I've gotta make it mergeable :)
