### 2024-02-05 7b18302
Clean up and refactor `KmerMinHash::merge` in core (#2973)
Use `Option::iter` to simplify code in `KmerMinhash::merge`, as well as
rewritting some checks so they are less redundant.

This wasn't quite the change I was going for, but I also noticed this
lowered Python tests runtime from 58s to 49s in my dev machine, mostly
because `similarity` is 17% faster, so might as well merge it.
