### 2024-02-05 075f780
disable rules generating errors in the ruff fix step
