+++ pyproject.toml
@@ -161,6 +161,7 @@ macos-deployment-target = "11.0"
 extend-select = [
   "UP",  # pyupgrade
 ]
+ignore = ["F401", "E712", "E402", "F821", "E722", "E741", "F811", "F403", "F822"]
 
 [tool.isort]
 known_third_party = ["deprecation", "hypothesis", "mmh3", "numpy", "pytest", "screed", "sourmash_tst_utils"]
