+++ .github/workflows/build_wheel_all_archs.yml
@@ -85,5 +85,5 @@ jobs:
       - name: Release
         uses: fnkr/github-action-ghr@v1
         env:
-          GHR_PATH: ${{steps.fetch_artifacts.outputs.download-path}}/artifact
+          GHR_PATH: ${{steps.fetch_artifacts.outputs.download-path}}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
