+++ src/sourmash/compare.py
@@ -39,22 +39,28 @@ def compare_serial(siglist, ignore_abundance, *, downsample=False, return_ani=Fa
 
     for i, j in iterator:
         if return_ani:
-            ani_result = siglist[i].jaccard_ani(siglist[j],downsample=downsample)
+            ani_result = siglist[i].jaccard_ani(siglist[j], downsample=downsample)
             if not potential_false_negatives and ani_result.p_exceeds_threshold:
                 potential_false_negatives = True
             if not jaccard_ani_untrustworthy and ani_result.je_exceeds_threshold:
                 jaccard_ani_untrustworthy = True
             ani = ani_result.ani
-            if ani == None:
+            if ani is None:
                 ani = 0.0
             similarities[i][j] = similarities[j][i] = ani
         else:
-            similarities[i][j] = similarities[j][i] = siglist[i].similarity(siglist[j], ignore_abundance=ignore_abundance, downsample=downsample)
+            similarities[i][j] = similarities[j][i] = siglist[i].similarity(
+                siglist[j], ignore_abundance=ignore_abundance, downsample=downsample
+            )
 
     if jaccard_ani_untrustworthy:
-        notify("WARNING: Jaccard estimation for at least one of these comparisons is likely inaccurate. Could not estimate ANI for these comparisons.")
+        notify(
+            "WARNING: Jaccard estimation for at least one of these comparisons is likely inaccurate. Could not estimate ANI for these comparisons."
+        )
     if potential_false_negatives:
-        notify("WARNING: Some of these sketches may have no hashes in common based on chance alone (false negatives). Consider decreasing your scaled value to prevent this.")
+        notify(
+            "WARNING: Some of these sketches may have no hashes in common based on chance alone (false negatives). Consider decreasing your scaled value to prevent this."
+        )
     return similarities
 
 
@@ -78,19 +84,24 @@ def compare_serial_containment(siglist, *, downsample=False, return_ani=False):
             if i == j:
                 containments[i][j] = 1
             elif return_ani:
-                ani_result = siglist[j].containment_ani(siglist[i], downsample=downsample)
+                ani_result = siglist[j].containment_ani(
+                    siglist[i], downsample=downsample
+                )
                 ani = ani_result.ani
                 if not potential_false_negatives and ani_result.p_exceeds_threshold:
                     potential_false_negatives = True
-                if ani == None:
+                if ani is None:
                     ani = 0.0
                 containments[i][j] = ani
             else:
-                containments[i][j] = siglist[j].contained_by(siglist[i],
-                                                         downsample=downsample)
+                containments[i][j] = siglist[j].contained_by(
+                    siglist[i], downsample=downsample
+                )
 
     if potential_false_negatives:
-        notify("WARNING: Some of these sketches may have no hashes in common based on chance alone (false negatives). Consider decreasing your scaled value to prevent this.")
+        notify(
+            "WARNING: Some of these sketches may have no hashes in common based on chance alone (false negatives). Consider decreasing your scaled value to prevent this."
+        )
 
     return containments
 
@@ -115,18 +126,23 @@ def compare_serial_max_containment(siglist, *, downsample=False, return_ani=Fals
 
     for i, j in iterator:
         if return_ani:
-            ani_result = siglist[j].max_containment_ani(siglist[i], downsample=downsample)
+            ani_result = siglist[j].max_containment_ani(
+                siglist[i], downsample=downsample
+            )
             ani = ani_result.ani
             if not potential_false_negatives and ani_result.p_exceeds_threshold:
                 potential_false_negatives = True
[…]
