+++ src/sourmash/sig/__main__.py
@@ -1,47 +1,55 @@
 """
 Command-line entry point for 'python -m sourmash.sig'
 """
-__all__ = ["cat",
-           "split",
-           "describe",
-           "manifest",
-           "overlap",
-           "merge",
-           "intersect",
-           "inflate",
-           "subtract",
-           "rename",
-           "extract",
-           "filter",
-           "flatten",
-           "downsample",
-           "ingest",
-           "export",
-           "kmers",
-           "fileinfo",
-           "check",
-           "collect"]
+__all__ = [
+    "cat",
+    "split",
+    "describe",
+    "manifest",
+    "overlap",
+    "merge",
+    "intersect",
+    "inflate",
+    "subtract",
+    "rename",
+    "extract",
+    "filter",
+    "flatten",
+    "downsample",
+    "ingest",
+    "export",
+    "kmers",
+    "fileinfo",
+    "check",
+    "collect",
+]
 
 import sys
 import csv
 import json
 import os
 from collections import defaultdict, namedtuple, Counter
-import json
 import re
 
 import screed
 import sourmash
 from sourmash.sourmash_args import FileOutput
 
-from sourmash.logging import (set_quiet, error, notify, print_results, debug,
-                              debug_literal, _debug)
+from sourmash.logging import (
+    set_quiet,
+    error,
+    notify,
+    print_results,
+    debug,
+    debug_literal,
+    _debug,
+)
 from sourmash import sourmash_args
 from sourmash.minhash import _get_max_hash_for_scaled
 from sourmash.manifest import CollectionManifest
 
 
-usage='''
+usage = """
 sourmash signature <command> [<args>] - manipulate/work with signature files.
 
 ** Commands can be:
@@ -67,15 +75,19 @@ collect <locations> -o manifest.sqlmf     - collect sigs metadata into manifest
 ** Use '-h' to get subcommand-specific help, e.g.
 
 sourmash signature merge -h
-'''
+"""
 
 
 def _check_abundance_compatibility(sig1, sig2):
     if sig1.minhash.track_abundance != sig2.minhash.track_abundance:
-        raise ValueError("incompatible signatures: track_abundance is {} in first sig, {} in second".format(sig1.minhash.track_abundance, sig2.minhash.track_abundance))
+        raise ValueError(
+            "incompatible signatures: track_abundance is {} in first sig, {} in second".format(
+                sig1.minhash.track_abundance, sig2.minhash.track_abundance
+            )
+        )
 
 
-def _extend_signatures_with_from_file(args, *, target_attr='signatures'):
+def _extend_signatures_with_from_file(args, *, target_attr="signatures"):
     # extend input signatures with --from-file
     if args.from_file:
         more_files = sourmash_args.load_pathlist_from_file(args.from_file)
@@ -109,7 +121,7 @@ def cat(args):
     picklist = sourmash_args.load_picklist(args)
     pattern_search = sourmash_args.load_include_exclude_db_patterns(args)
 
-    encountered_md5sums = defaultdict(int)   # used by --unique
+    encountered_md5sums = defaultdict(int)  # used by --unique
 
     # open output for saving sigs
     save_sigs = sourmash_args.SaveSignaturesToLocation(args.output)
@@ -119,14 +131,16 @@ def cat(args):
 
     # start loading!
     progress = sourmash_args.SignatureLoadingProgress()
-    loader = sourmash_args.load_many_signatures(args.signatures,
-                                                ksize=args.ksize,
-                                                moltype=moltype,
-                                                picklist=picklist,
-                                                progress=progress,
-                                                yield_all_files=args.force,
-                                                force=args.force,
-                                                pattern=pattern_search)
+    loader = sourmash_args.load_many_signatures(
+        args.signatures,
+        ksize=args.ksize,
+        moltype=moltype,
[…]
