### 2024-02-05 5e6fdb9
MRG: update the CLI docs and help for `search --containment` and `prefetch` (#2971)
Adds useful information about the order of containment searches:
* `search --containment A B` reports A contained in B;
* `prefetch A B` reports B contained in A;

Fixes https://github.com/sourmash-bio/sourmash/issues/2968.
