### 2024-02-05 2f286fb
update CI to run on macos-arm64 too, with nix-installer
