### 2024-02-05 b265415
MRG: fix upload wheel CI (#2974)
Update paths for upload of wheels on release, based on
https://github.com/sourmash-bio/sourmash/pull/2887.
