+++ .github/workflows/build_wheel_all_archs.yml
@@ -1,4 +1,5 @@
 name: maturin wheels
+permissions: write-all
 
 on:
   pull_request:        # use for testing modifications to this action
@@ -8,9 +9,6 @@ on:
   workflow_dispatch:
   schedule:
     - cron: "0 0 * * *" # daily
-
-permissions:
-  contents: read
 
 jobs:
 
@@ -85,5 +83,5 @@ jobs:
       - name: Release
         uses: fnkr/github-action-ghr@v1
         env:
-          GHR_PATH: ${{steps.fetch_artifacts.outputs.download-path}}/artifact
+          GHR_PATH: ${{steps.fetch_artifacts.outputs.download-path}}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
