+++ .github/workflows/build_wheel.yml
@@ -1,4 +1,5 @@
 name: cibuildwheel
+permissions: write-all
 
 on:
   push:
@@ -104,5 +105,5 @@ jobs:
       - name: Release
         uses: fnkr/github-action-ghr@v1
         env:
-          GHR_PATH: ${{steps.fetch_artifacts.outputs.download-path}}/artifact
+          GHR_PATH: ${{steps.fetch_artifacts.outputs.download-path}}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
