### 2024-02-05 fee6292
Pre-commit updates (#2427)
Ref #2421 
Fix #2908

Updates older pre-commit hooks (initial exp:
https://github.com/sourmash-bio/sourmash/pull/680), mostly based (again)
on [tox
configs](https://github.com/tox-dev/tox/blob/main/.pre-commit-config.yaml)

Use ruff instead of pyupgrade/isort/black/flake8

This PR has a couple of commits: the first ones update configs, the last
one runs `tox -e fix_lint` to apply pre-commit. Mostly looking into
updating the first commits without breaking tests on the last commit.
