+++ src/sourmash/cli/tax/annotate.py
@@ -1,6 +1,6 @@
 """annotate gather results with taxonomy information"""
 
-usage="""
+usage = """
 
     sourmash tax annotate --gather-csv <gather_csv> [ ... ] --taxonomy-csv <taxonomy_csv> [ ... ]
 
@@ -19,53 +19,70 @@ from sourmash.logging import notify, print_results, error
 
 
 def subparser(subparsers):
-    subparser = subparsers.add_parser('annotate',
-                                      aliases=['annotate'],
-                                      usage=usage)
+    subparser = subparsers.add_parser("annotate", aliases=["annotate"], usage=usage)
     subparser.add_argument(
-        '-g', '--gather-csv', nargs='*', default = [], action='extend',
-        help='CSV output files from sourmash gather'
+        "-g",
+        "--gather-csv",
+        nargs="*",
+        default=[],
+        action="extend",
+        help="CSV output files from sourmash gather",
     )
     subparser.add_argument(
-        '--from-file',  metavar='FILE', default=None,
-        help='input many gather results as a text file, with one gather CSV per line'
+        "--from-file",
+        metavar="FILE",
+        default=None,
+        help="input many gather results as a text file, with one gather CSV per line",
     )
     subparser.add_argument(
-        '-q', '--quiet', action='store_true',
-        help='suppress non-error output'
+        "-q", "--quiet", action="store_true", help="suppress non-error output"
     )
     subparser.add_argument(
-        '-t', '--taxonomy-csv', '--taxonomy', metavar='FILE',
-        nargs='*', required=True, action="extend",
-        help='database lineages CSV'
+        "-t",
+        "--taxonomy-csv",
+        "--taxonomy",
+        metavar="FILE",
+        nargs="*",
+        required=True,
+        action="extend",
+        help="database lineages CSV",
     )
     subparser.add_argument(
-        '-o', '--output-dir', default= "",
-        help='directory for output files'
+        "-o", "--output-dir", default="", help="directory for output files"
     )
     subparser.add_argument(
-        '--keep-full-identifiers', action='store_true',
-        help='do not split identifiers on whitespace'
+        "--keep-full-identifiers",
+        action="store_true",
+        help="do not split identifiers on whitespace",
     )
     subparser.add_argument(
-        '--keep-identifier-versions', action='store_true',
-        help='after splitting identifiers, do not remove accession versions'
+        "--keep-identifier-versions",
+        action="store_true",
+        help="after splitting identifiers, do not remove accession versions",
     )
     subparser.add_argument(
-        '--fail-on-missing-taxonomy', action='store_true',
-        help='fail quickly if taxonomy is not available for an identifier',
+        "--fail-on-missing-taxonomy",
+        action="store_true",
+        help="fail quickly if taxonomy is not available for an identifier",
     )
     subparser.add_argument(
-        '-f', '--force', action = 'store_true',
-        help='continue past errors in file and taxonomy loading',
+        "-f",
+        "--force",
+        action="store_true",
+        help="continue past errors in file and taxonomy loading",
     )
     subparser.add_argument(
-        '--lins', '--lin-taxonomy', action='store_true', default=False,
-        help='use LIN taxonomy in place of standard taxonomic ranks.  Note that the taxonomy CSV must contain LIN lineage information.'
+        "--lins",
+        "--lin-taxonomy",
+        action="store_true",
+        default=False,
+        help="use LIN taxonomy in place of standard taxonomic ranks.  Note that the taxonomy CSV must contain LIN lineage information.",
     )
 
+
 def main(args):
-    import sourmash
     if not args.gather_csv and not args.from_file:
[…]
