+++ tests/test_prefetch.py
@@ -25,29 +25,40 @@ def test_prefetch_basic(runtmp, linear_gather):
     c = runtmp
 
     # test a basic prefetch
-    sig2 = utils.get_test_data('2.fa.sig')
-    sig47 = utils.get_test_data('47.fa.sig')
-    sig63 = utils.get_test_data('63.fa.sig')
+    sig2 = utils.get_test_data("2.fa.sig")
+    sig47 = utils.get_test_data("47.fa.sig")
+    sig63 = utils.get_test_data("63.fa.sig")
 
-    c.run_sourmash('prefetch', '-k', '31', sig47, sig63, sig2, sig47,
-                   linear_gather)
+    c.run_sourmash("prefetch", "-k", "31", sig47, sig63, sig2, sig47, linear_gather)
     print(c.last_result.status)
     print(c.last_result.out)
     print(c.last_result.err)
 
     assert c.last_result.status == 0
 
-    assert "WARNING: no output(s) specified! Nothing will be saved from this prefetch!" in c.last_result.err
+    assert (
+        "WARNING: no output(s) specified! Nothing will be saved from this prefetch!"
+        in c.last_result.err
+    )
     assert "selecting specified query k=31" in c.last_result.err
-    assert "loaded query: NC_009665.1 Shewanella baltica... (k=31, DNA)" in c.last_result.err
-    assert "query sketch has scaled=1000; will be dynamically downsampled as needed" in c.last_result.err
+    assert (
+        "loaded query: NC_009665.1 Shewanella baltica... (k=31, DNA)"
+        in c.last_result.err
+    )
+    assert (
+        "query sketch has scaled=1000; will be dynamically downsampled as needed"
+        in c.last_result.err
+    )
 
     err = c.last_result.err
     assert "loaded 5 total signatures from 3 locations." in err
     assert "after selecting signatures compatible with search, 3 remain." in err
 
     assert "total of 2 matching signatures." in c.last_result.err
-    assert "of 5177 distinct query hashes, 5177 were found in matches above threshold." in c.last_result.err
+    assert (
+        "of 5177 distinct query hashes, 5177 were found in matches above threshold."
+        in c.last_result.err
+    )
     assert "a total of 0 query hashes remain unmatched." in c.last_result.err
 
 
@@ -55,15 +66,18 @@ def test_prefetch_select_query_ksize(runtmp, linear_gather):
     # test prefetch where query and subject db both have multiple ksizes
     c = runtmp
 
-    ss = utils.get_test_data('GCF_000005845.2_ASM584v2_genomic.fna.gz.sig')
+    ss = utils.get_test_data("GCF_000005845.2_ASM584v2_genomic.fna.gz.sig")
 
-    c.run_sourmash('prefetch', ss, ss, linear_gather)
+    c.run_sourmash("prefetch", ss, ss, linear_gather)
     print(c.last_result.status)
     print(c.last_result.out)
     print(c.last_result.err)
 
     assert c.last_result.status == 0
-    assert 'of 4476 distinct query hashes, 4476 were found in matches above threshold.' in c.last_result.err
+    assert (
+        "of 4476 distinct query hashes, 4476 were found in matches above threshold."
+        in c.last_result.err
+    )
 
 
 def test_prefetch_subject_scaled_is_larger(runtmp, linear_gather):
@@ -71,26 +85,39 @@ def test_prefetch_subject_scaled_is_larger(runtmp, linear_gather):
     c = runtmp
 
     # make a query sketch with scaled=1000
-    fa = utils.get_test_data('genome-s10.fa.gz')
-    c.run_sourmash('sketch', 'dna', fa, '-o', 'query.sig')
[…]
