+++ tests/conftest.py
@@ -5,10 +5,13 @@ from hypothesis import settings, Verbosity
 import pytest
 
 import matplotlib.pyplot as plt
-plt.rcParams.update({'figure.max_open_warning': 0})
+
+plt.rcParams.update({"figure.max_open_warning": 0})
 
 from sourmash_tst_utils import TempDirectory, RunnerContext
+
 sys.stdout = sys.stderr
+
 
 @pytest.fixture
 def runtmp():
@@ -66,16 +69,17 @@ def use_manifest(request):
     return request.param
 
 
-@pytest.fixture(params=['json', 'sql'])
+@pytest.fixture(params=["json", "sql"])
 def lca_db_format(request):
     return request.param
 
 
-@pytest.fixture(params=['csv', 'sql'])
+@pytest.fixture(params=["csv", "sql"])
 def manifest_db_format(request):
     return request.param
 
-@pytest.fixture(params=['sig', 'sig.gz', 'zip', '.d/', '.sqldb'])
+
+@pytest.fixture(params=["sig", "sig.gz", "zip", ".d/", ".sqldb"])
 def sig_save_extension(request):
     return request.param
 
@@ -89,29 +93,37 @@ def pytest_collection_modifyitems(items, config):
         deselected_items = []
 
         for item in items:
-            if fixture_name in getattr(item, 'fixturenames', ()):
+            if fixture_name in getattr(item, "fixturenames", ()):
                 selected_items.append(item)
             else:
                 deselected_items.append(item)
         config.hook.pytest_deselected(items=deselected_items)
         items[:] = selected_items
+
+
 # --- END - Only run tests using a particular fixture --- #
 
-def pytest_addoption(parser):
-    parser.addoption("--usesfixture",
-                     action="store",
-                     default=None,
-                     help="just run tests that use a particular fixture")
 
-    parser.addoption("--run-hypothesis", action="store_true",
-                     help="run hypothesis tests")
+def pytest_addoption(parser):
+    parser.addoption(
+        "--usesfixture",
+        action="store",
+        default=None,
+        help="just run tests that use a particular fixture",
+    )
+
+    parser.addoption(
+        "--run-hypothesis", action="store_true", help="run hypothesis tests"
+    )
+
 
 def pytest_runtest_setup(item):
     if item.config.getoption("--run-hypothesis"):
         if not any(mark for mark in item.iter_markers(name="hypothesis")):
             pytest.skip("--run-hypothesis option set, running only hypothesis tests")
 
+
 settings.register_profile("ci", max_examples=1000)
 settings.register_profile("dev", max_examples=10)
 settings.register_profile("debug", max_examples=10, verbosity=Verbosity.verbose)
-settings.load_profile(os.getenv(u'HYPOTHESIS_PROFILE', 'default'))
+settings.load_profile(os.getenv("HYPOTHESIS_PROFILE", "default"))

[…]
