### 2024-02-05 3ce9baf
fix syntax for running on linux and macos on CI
