### 2024-02-05 8394a77
fix nix command to run
