### 2024-02-06 c6831fd
MRG: add label output & input options to `compare` and `plot`, for better customization (#2598)
Adds `sourmash compare --labels-to` and `sourmash plot --labels-from` to
support better label customization.

Fixes https://github.com/sourmash-bio/sourmash/issues/2452
Fixes https://github.com/sourmash-bio/sourmash/issues/2915

## `sourmash compare --labels-to`

This command will generate a 'labels-to' file. Running:
```
sourmash compare tests/test-data/demo/*.sig -o compare-demo \
    --labels-to compare-demo-labels.csv
```
will produce a file that looks like this:

file `compare-demo-labels.csv`:
[…]
