+++ src/sourmash/cli/compare.py
@@ -95,6 +95,11 @@ def subparser(subparsers):
         help="write matrix to specified file in CSV format (with column " "headers)",
     )
     subparser.add_argument(
+        "--labels-to",
+        "--labels-save",
+        help="a CSV file containing label information",
+    )
+    subparser.add_argument(
         "-p",
         "--processes",
         metavar="N",
