+++ tests/test_sourmash.py
@@ -151,6 +151,7 @@ def test_compare_serial(runtmp):
 
     testsigs = utils.get_test_data("genome-s1*.sig")
     testsigs = glob.glob(testsigs)
+    assert len(testsigs) == 4
 
     c.run_sourmash("compare", "-o", "cmp", "-k", "21", "--dna", *testsigs)
 
@@ -1252,7 +1253,7 @@ def test_plot_override_labeltext(runtmp):
 
     print(runtmp.last_result.out)
 
-    assert "loading labels from new.labels.txt" in runtmp.last_result.err
+    assert "loading labels from text file 'new.labels.txt'" in runtmp.last_result.err
 
     expected = """\
 0\ta
@@ -1291,7 +1292,7 @@ def test_plot_override_labeltext_fail(runtmp):
     print(runtmp.last_result.out)
     print(runtmp.last_result.err)
     assert runtmp.last_result.status != 0
-    assert "loading labels from new.labels.txt" in runtmp.last_result.err
+    assert "loading labels from text file 'new.labels.txt'" in runtmp.last_result.err
     assert "3 labels != matrix size, exiting" in runtmp.last_result.err
 
 
@@ -1404,6 +1405,117 @@ def test_plot_subsample_2(runtmp):
 1\tgenome-s10+s11
 2\tgenome-s11"""
     assert expected in runtmp.last_result.out
+
+
+def test_compare_and_plot_labels_from_to(runtmp):
+    # test doing compare --labels-to and plot --labels-from.
+    testdata1 = utils.get_test_data("genome-s10.fa.gz.sig")
+    testdata2 = utils.get_test_data("genome-s11.fa.gz.sig")
+    testdata3 = utils.get_test_data("genome-s12.fa.gz.sig")
+    testdata4 = utils.get_test_data("genome-s10+s11.sig")
+
+    labels_csv = runtmp.output("label.csv")
+
+    runtmp.run_sourmash(
+        "compare",
+        testdata1,
+        testdata2,
+        testdata3,
+        testdata4,
+        "-o",
+        "cmp",
+        "-k",
+        "21",
+        "--dna",
+        "--labels-to",
+        labels_csv,
+    )
+
+    runtmp.sourmash("plot", "cmp", "--labels-from", labels_csv)
+
+    print(runtmp.last_result.out)
+
+    assert "loading labels from CSV file" in runtmp.last_result.err
+
+    expected = """\
+0\tgenome-s10
+1\tgenome-s11
+2\tgenome-s12
+3\tgenome-s10+s11"""
+    assert expected in runtmp.last_result.out
+
+
+def test_compare_and_plot_labels_from_changed(runtmp):
+    # test 'plot --labels-from' with changed labels
+    testdata1 = utils.get_test_data("genome-s10.fa.gz.sig")
+    testdata2 = utils.get_test_data("genome-s11.fa.gz.sig")
+    testdata3 = utils.get_test_data("genome-s12.fa.gz.sig")
+    testdata4 = utils.get_test_data("genome-s10+s11.sig")
+
+    labels_csv = utils.get_test_data("compare/labels_from-test.csv")
+
+    runtmp.run_sourmash(
+        "compare",
+        testdata1,
+        testdata2,
+        testdata3,
+        testdata4,
+        "-o",
+        "cmp",
+        "-k",
+        "21",
+        "--dna",
+    )
+
+    runtmp.sourmash("plot", "cmp", "--labels-from", labels_csv)
+
+    print(runtmp.last_result.out)
+
+    assert "loading labels from CSV file" in runtmp.last_result.err
+
+    expected = """\
+0\tgenome-s10-CHANGED
+1\tgenome-s11-CHANGED
+2\tgenome-s12-CHANGED
[…]
