### 2024-02-10 7edcf14
nix updates for pyopensci review (#2975)
Fix #2905 #2907 

- Use `DeterminateSystems/nix-installer-action` and
`DeterminateSystems/magic-nix-cache-action` for installing nix and
caching
- Also run nix on macos-14 (arm64)
- run `nix develop` instead of `nix-shell` to mirror direnv
