### 2024-02-11 ec898ff
Fix tox and nix configs so all tox tests execute correctly (#2992)
Closes #2907

`tox -e docs` was failing due to missing env vars in tox.ini config
`tox -e khmer_master` fails to due to external errors now, it was
missing the URL for installing latest version from github.

Also limit test discovery in `hypothesis`, `khmer`, and `khmer_master`
to `doc tests`, avoiding going into `.tox` and other internal
directories containing python code.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
