### 2024-02-12 sourmash-bio/sourmash#2890 @ctb
our response! https://github.com/pyOpenSci/software-submission/issues/129#issuecomment-1939369
875
