### 2024-04-04 #129 @snacktavish
🎉Sourmash has been approved by pyOpenSci! Thank you @bluegenes @luizirber  and @ctb for submitting sourmash and many thanks to @elais and @LilyAnderssonLee  for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

---
