### 2024-04-06 #129 @ctb
thanks all! really happy! :tada:

question: if we want to go for JOSS, should we do a new release _now_, for pyopensci, and then _another_, assuming JOSS accepts? or can we wait and just do a release if/when JOSS accepts? we don't have a real reason for a new release just yet (we've been doing them regularly anyway!) so could go either way. Just curious.

(we have citation instructions in the software that we would change for the JOSS release.)
