### 2024-04-06 20e6195
MRG: add pyopensci review badge (#3105)
Add pyopensci review badge, linking to accepted review:
https://github.com/pyOpenSci/software-submission/issues/129
