### 2024-04-08 #129 @snacktavish
I checked in with @lwasser and we say -
Do a release now for pyos then another bump for the JOSS paper which than can be linked to the DOI that joss provides!

Thank you!
