### 2024-04-09 #129 @bluegenes
Thanks @snacktavish @lwasser!!

for tracking:

# Author Wrap Up Tasks

 - [x] Tag and create a release to create a Zenodo version and DOI.  [(release 4.8.8)](https://github.com/sourmash-bio/sourmash/releases/tag/v4.8.8) - https://zenodo.org/records/10951577
 - [x] Add the badge for pyOpenSci peer-review to the README.md of . The badge should be [![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number).

- [ ] Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.
- [ ] Login to the JOSS website and fill out the JOSS submission form using your Zenodo DOI. When you fill out the form, be sure to mention and link to the approved pyOpenSci review. JOSS will tag your package for expedited review if it is already pyOpenSci approved.
