### 2024-06-06 #129 @lwasser
hey there everyone! i'm just checking in on editor wrap up tasks here!! a few things

@bluegenes @LilyAnderssonLee @elais - if you have time, can you kindly complete our [post-review survey](https://docs.google.com/forms/d/e/1FAIpQLSce0vICvRBuMr-7Oq9hsiemn46RmQdS4iq4Y_iya8eu5Jh6mQ/viewform)? we really appreciate your input as it helps us improve our review process and it also helps us when we write writeups for funders, etc demonstrate impact. It looks like titus did complete it

* [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10951577.svg)](https://doi.org/10.5281/zenodo.10951577)
* date accepted 4 April 2024
* i have just pinged the JOSS editor to let them know this is a fast track. @bluegenes can you please let me know if you see the package moving towards a full review again? this step should be super fast. they should only look at your paper and accept our review as theirs. So the only reason why it wouldn't be accepted is that this is a second review so you have to demonstrate significant updates to the package (i suspect there are because of that @ctb has mentioned in slack). but this is JOSS' call not ours!
