### 2024-06-06 #129 @lwasser
Ok a few more items

@bluegenes @ctb

We'd LOVE for you and your team to write a blog post on SourMash for us to promote your work! if you are interested - here are a few examples of other blog posts:

* [sleplet](https://www.pyopensci.org/blog/sleplet-slepian-wavelets)
* [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
* [movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)

you can direct any questions to @kierisi on the blog post if that is of interest!

 [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

it can even be a tutorial like post that highlights what your package does. Then we can share it with people to get the word out about your package.

If you are too busy for this no worries. But if you have time - we'd love to spread the word about your package!

Finally - EVERYONE involved in this review (authors, editors, reviewers, etc) is invited to join our pyOpenSci slack. If you are interested, please reply and @kierisi can ensure that you get an invite! Thank you everyone. You may here from me again as i just cross additional  `t's` and dot `i's` in these last steps of the review! 🚀

I really appreciate ALL of the effort htat went into this review from EVERYONE here in this issue. Thank you all for being a part of the wonderful pyOpenSci community. You all make it the supportive community that it is 💟
