### 2024-06-06 #129 @bluegenes
> We'd LOVE for you and your team to write a blog post on SourMash for us to promote your work!

we'd love to -- I'll reach out again when we have something.

> Finally - EVERYONE involved in this review (authors, editors, reviewers, etc) is invited to join our pyOpenSci slack. If you are interested, please reply and @kierisi can ensure that you get an invite! Thank you everyone. You may here from me again as i just cross additional t's and dot i's in these last steps of the review! 🚀

yes - I'd be happy to join the slack!
